% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{AcousticPSU}
\alias{AcousticPSU}
\title{Add or remove biotic stations from assignments}
\description{
The functions \code{addStations} and \code{removeStations} adds or removes biotic stations from the Assignment process data of the specified process.
}
\details{
The assignment IDs are refreshed for every change, after sorting the assignemnts by the PSU column.
}
