% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{addProcess}
\alias{addProcess}
\title{Add a StoX process to a model.}
\usage{
addProcess(
  projectPath,
  modelName,
  values = NULL,
  returnProcessTable = TRUE,
  archive = TRUE,
  add.defaults = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{values}{A list of values to assign to the process, such as list(processName = "ReadBiotic", functionName = "RstoxBase::ReadBiotic").}

\item{returnProcessTable}{Logical: If TRUE return the process table.}
}
\description{
Add a StoX process to a model.
}
