% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{BioticAssignment}
\alias{BioticAssignment}
\alias{addHaulToAssignment}
\alias{removeHaulFromAssignment}
\title{Add or remove biotic hauls to/from assignment of hauls to acoustic PSU/Layer}
\usage{
addHaulToAssignment(Stratum, PSU, Haul, projectPath, modelName, processID)

removeHaulFromAssignment(Stratum, PSU, Haul, projectPath, modelName, processID)
}
\arguments{
\item{Stratum}{The stratum name related to the PSU.}

\item{PSU}{The acoustic primary sampling unit (PSU) for which to add/remove the haul.}

\item{Haul}{The biotic haul to add/remove (can be a vector of hauls).}

\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}
}
\description{
The functions \code{addHaulsToAssignment} and \code{removeHaulsFromAssignment} adds or removes biotic hauls from the \code{\link{BioticAssignment}} process data of the specified process.
}
\details{
The assignment IDs are refreshed for every change, after sorting the assignemnts by the PSU column.
}
