% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{BioticPSU}
\alias{BioticPSU}
\alias{addBioticPSU}
\alias{removeBioticPSU}
\alias{addStation}
\alias{removeStation}
\title{Add, remove and renamme biotic PSUs and Stations}
\usage{
addBioticPSU(Stratum, PSU = NULL, projectPath, modelName, processID)

removeBioticPSU(PSU, projectPath, modelName, processID)

addStation(PSU, Station, projectPath, modelName, processID)

removeStation(Station, projectPath, modelName, processID)
}
\arguments{
\item{Stratum}{The name of a stratum.}

\item{PSU}{The name of a PSU.}

\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{Station}{The name of a biotic station.}
}
\description{
The functions \code{addBioticPSU}, \code{removeBioticPSU} and \code{renameBioticPSU} adds, removes or renames BioticPSUs from the BioticPSU process data of the specified process. 
The functions \code{addStation} and \code{removeStation} adds or removes Stations from the BioticPSU process data of the specified process. 
The function \code{removeAllBioticPSUsOfStratum} removes all PSUs from the BioticPSU process data of the specified process.
}
\details{
The assignment IDs are refreshed for every change, after sorting the assignemnts by the PSU column.
}
