% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{BootstrapNetCDF4Data}
\alias{BootstrapNetCDF4Data}
\title{BootstrapNetCDF4 data}
\description{
The BootstrapNetCDF4Data model data is the path to a NetCDF4 memory file holding the outputs from the baseline generated by re-running baseline with random sampling of the PSUs and assignments (in the case of acoustic-trawl estimates).
}
\seealso{
This data type is produced by the function \code{\link{BootstrapNetCDF4}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxData}}, \code{\link{RstoxBase}} and \code{\link{RstoxFramework}}.
}
