% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{bootstrapRDataToNetCDF4}
\alias{bootstrapRDataToNetCDF4}
\title{Convert a bootstrap RData file to NetCDF4}
\usage{
bootstrapRDataToNetCDF4(
  bootstrapRDataFile,
  bootstrapNetCDF4File = file.path(dirname(bootstrapRDataFile), "BootstrapData.nc"),
  outputVariables = NULL,
  ow = FALSE
)
}
\arguments{
\item{bootstrapRDataFile}{The path to file holding the BoostrapData from a StoX <= 3.6.2 project run.}

\item{bootstrapNetCDF4File}{The path to the new netCDF4 file.}

\item{outputVariables}{An optional list of variables to keep in the output. A typical set of variables could be ["Survey", "Stratum", "SpeciesCategory", "IndividualTotalLength", "IndividualAge", "Abundance", "Biomass"], which should cover the most frequently used variables in reports. Any variable that is used in a report must be present in \code{outputVariables}. Empty list (the default) implies to keep all variables. This parameter is included to facilitate smaller disc space for the bootstrap objects and faster writing/reading of that file.}

\item{ow}{Logical: If TRUE, overwrite the \code{bootstrapNetCDF4File}.}
}
\description{
Convert a bootstrap RData file to NetCDF4
}
