% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{compareProjectToStoredOutputFilesAll}
\alias{compareProjectToStoredOutputFilesAll}
\title{Function for comparing existing output files with the memory read using runProject()}
\usage{
compareProjectToStoredOutputFilesAll(
  projectPaths,
  projectPaths_original = projectPaths,
  emptyStringAsNA = FALSE,
  intersect.names = TRUE,
  ignore.variable = NULL,
  skipNAFraction = FALSE,
  skipNAAt = FALSE,
  NAReplacement = NULL,
  classOf = c("first", "second"),
  try = TRUE,
  data.out = FALSE,
  ...
)
}
\arguments{
\item{projectPaths}{The projects to be run and tested against the existing output files of the project gievn by \code{projectPath_original}.}

\item{projectPaths_original}{The projects holding the existing output files, defaulted to \code{projectPath}.}

\item{emptyStringAsNA}{Logical: If TRUE, read empty strings as NA from the stored original tables, as RstoxFramework has started writing NAs as NAs and not as empty strings.}

\item{intersect.names}{Logical: If TRUE, compare only same named columns.}

\item{ignore.variable}{A vector of names of variables/columns to ignore in the comparison.}

\item{skipNAFraction}{Logical: If TRUE, skip rows with more than 50 percent NAs. Can be set to a value between 0 and 1.}

\item{skipNAAt}{A vector of strings naming the columns in which NA values identifies rows to skip.}

\item{NAReplacement}{List of replacement values for different classes of NA, applied after any merging as to incorporate NAs generated during merging.}

\item{classOf}{Character string specifying whether to compare after converting to the class of the first or second table. Set this to "first" (default) to convert class to the original data.}

\item{try}{Logical: If FALSE do not run the process in a \code{tryCatch}. Set this to FALSE when debugging, as the \code{tryCatch} masks the errors in the \code{traceback}.}

\item{data.out}{Logical, if TRUE output the original and new data along with the tests. \code{data.out} = NULL implies \code{data.out} = FALSE if no difference was found and  \code{data.out} = TRUE otherwise.}

\item{...}{\code{replaceArgsList} can also be given directly.}
}
\description{
Function for comparing existing output files with the memory read using runProject()
}
