% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{compareSweptAreaBaseline}
\alias{compareSweptAreaBaseline}
\title{Compare processes in a StoX 2.7 project and a copy in StoX >= 3.0.0:}
\usage{
compareSweptAreaBaseline(
  projectPathOld,
  projectPathNew,
  comparisonModel,
  output_Old = NULL,
  output_New = NULL,
  data.out = FALSE
)
}
\arguments{
\item{projectPathOld}{The path to the StoX 2.7 project.}

\item{projectPathNew}{The path to the copy project in StoX >= 3.0.0.}

\item{comparisonModel}{A list with specifications of what and how to compare the outputs from the old and new project (see Details).}

\item{output_Old}{The StoX 2.7 table.}

\item{output_New}{The StoX 2.7 table.}

\item{data.out}{The StoX 2.7 table.

The \code{comparisonModel} is a list with the following required elements:
\itemize{
 \item{dataType}{The StoX >= 3.0.0 datatypes to compare. Used only for reference.}
 \item{dataVariable_Old}{The specific variable to compare in the old datatypes.}
 \item{dataVariable_New}{The specific variable to compare in the new datatypes.}
 \item{processName_Old,processName_New}{The process named in the old and new project, respectively. Must correspond to the \code{dataType}.}
 \item{compareAction}{A string vector specifyfing whether to merge ("merge") or cbind ("cbind") when comparing. Cbind is only for SuperIndividuals in StoX 2.7, which does not have unique keys in.}
 \item{keys_Old,keys_New"}{A list of the keys of the datatypes in the old and new project, respectively. Specify only for datatypes to be merged (use NA for those that use compareAction = "cbind").}
 \item{subsetByNAOn_Old,subsetByNAOn_New"}{A list of variables to subset the old and new data by, respectively. When any of these are NA in a row, the row is skipped. Specify only for datatypes to be cbinded (use NA for those that use compareAction = "merge").}
 \item{orderBy_Old,orderBy_New}{SA list of variables to order the old and new data by, respectively (after subsetting using \code{subsetByNAOn_Old} and \code{subsetByNAOn_New}). Specify only for datatypes to be cbinded (use NA for those that use compareAction = "merge").}
 \item{tolerance}{A numeric vector setting the tolerance for the comparison, below which the outputs are regarded as equal.}
}}
}
\description{
Compare processes in a StoX 2.7 project and a copy in StoX >= 3.0.0:
}
