% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{convertKeys2.7To3}
\alias{convertKeys2.7To3}
\title{Convert keys of a 2.7 table to mimic the keys of StoX >= 3}
\usage{
convertKeys2.7To3(x)
}
\arguments{
\item{x}{The StoX 2.7 table.}
}
\description{
Convert keys of a 2.7 table to mimic the keys of StoX >= 3
}
