% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{download_files_ftp}
\alias{download_files_ftp}
\alias{list_files_ftp}
\title{Download files ftom and FTP sever, possibly recursively in the folder structure of the server}
\usage{
download_files_ftp(ftp, dir, quiet = TRUE, recursive = FALSE)

list_files_ftp(ftp, quiet = TRUE, recursive = FALSE)
}
\arguments{
\item{ftp}{The path to an ftp server.}

\item{dir}{The path to the directory in which to place the files.}

\item{quiet}{Logical: If TRUE suppress download messages.}

\item{recursive}{Logical: If TRUE recurse into sub folders.}
}
\value{
A data frame the columns \code{path} and \code{isDir}.
}
\description{
The function \code{download_files_ftp} downloads files form an FTP srever using the function \code{download_files_ftp} which list files and directories.
}
