% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{downloadStoxReference}
\alias{downloadStoxReference}
\title{Download StoX reference files}
\usage{
downloadStoxReference(ftp = "ftp://ftp.imr.no/StoX/Download/reference/",
  quiet = TRUE)
}
\arguments{
\item{ftp}{The path to an ftp server.}

\item{quiet}{Logical: If TRUE suppress download messages.}
}
\value{
A data frame the columns \code{path} and \code{isDir}.
}
\description{
This function downloads reference files from the StoX ftp server.
}
