% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{duplicateProcess}
\alias{duplicateProcess}
\title{Duplicate a StoX process.}
\usage{
duplicateProcess(projectPath, modelName, processID, newProcessName = NULL)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{newProcessName}{The name of the new process. The default is the name of the process to copy added "_copy".}
}
\description{
Duplicate a StoX process.
}
