% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{expandDT}
\alias{expandDT}
\title{Function to expand a data table so that the cells that are vectors are transposed and the rest repeated to fill the gaps}
\usage{
expandDT(DT, toExpand = NULL)
}
\arguments{
\item{DT}{A data.table.}

\item{toExpand}{A vector of names of the tables to expand.}
}
\description{
Function to expand a data table so that the cells that are vectors are transposed and the rest repeated to fill the gaps
}
