% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{expandProcess}
\alias{expandProcess}
\title{Insert a StoX process after a specific process}
\usage{
expandProcess(
  projectPath,
  modelName,
  processName,
  values = NULL,
  returnProcessTable = TRUE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processName}{The name of the process.}

\item{values}{A list of values to assign to the process, such as list(processName = "ReadBiotic", functionName = "RstoxBase::ReadBiotic").}

\item{returnProcessTable}{Logical: If TRUE return the process table (much used in functions used by the GUI).}
}
\description{
Insert a StoX process after a specific process
}
