% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{expression2list}
\alias{expression2list}
\title{Parse an R expression to a nested list:}
\usage{
expression2list(expr, generateRuleset = TRUE)
}
\arguments{
\item{expr}{An R expression string such as "a == 3 && !(b > 0)".}

\item{generateRuleset}{Logical: If TRUE output the expressions as rules, which are lists of condition and rules.}
}
\description{
Parse an R expression to a nested list:
}
