% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{extractPackageNameAsNames}
\alias{extractPackageNameAsNames}
\title{Small function to parse the string defining officical Rstox-package versions (for each RstoxFramwork):}
\usage{
extractPackageNameAsNames(packageVersionString)
}
\arguments{
\item{packageVersionString}{Character: A vector of strings of the form PACKAGENAME_PACKAGEVERSION (e.g. "RstoxFramework_3.6.1") to be converted into a list of versions named by the package names.}
}
\description{
Small function to parse the string defining officical Rstox-package versions (for each RstoxFramwork):
}
