% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{fixedWidthTable}
\alias{fixedWidthTable}
\title{Function to convert data.table to fixed width:}
\usage{
fixedWidthTable(
  x,
  columnSeparator = " ",
  lineSeparator = NULL,
  na = "-",
  enable.auto_unbox = TRUE
)
}
\arguments{
\item{x}{The table to modify.}

\item{columnSeparator}{The string to separate columns by, defaulted to a single space.}

\item{lineSeparator}{The string to separate lines by, defaulted to a NULL, which keeps the output as a vector of strings.}

\item{na}{The string to replace NAs by, defaulted to "-".}

\item{enable.auto_unbox}{Logical: If TRUE wrap the output in a list if  \code{pretty} is TRUE and the output is of length 1. This keeps the array when converting to JSON also for length 1.}
}
\description{
Function to convert data.table to fixed width:
}
