% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{general_arguments}
\alias{general_arguments}
\title{General parameters of RstoxFramework.}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{processName}{The name of the process.}

\item{tableName}{The name of the output table to get from the process.}

\item{template}{A string naming the template to use when generating the project. See \code{getAvaiableTemplates} for a list of available templates.}

\item{ow}{Logical: If TRUE overwrite the project.}

\item{showWarnings}{Logical: If TRUE display warninigs when creting the project folders.}

\item{Application}{A string naming the application (e.g. graphical user interface such as StoX) used to produce the description file.}
}
\description{
All functions referring to a project, a model, a process or an output table use the same parameters, listed here.
}
