% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes.R
\name{getAllParameterTableInfo}
\alias{getAllParameterTableInfo}
\title{GUI function: Get a table of all parameter table info, holding tha name and type of all columns defined in parameter tables in all Rstox packages}
\usage{
getAllParameterTableInfo()
}
\description{
GUI function: Get a table of all parameter table info, holding tha name and type of all columns defined in parameter tables in all Rstox packages
}
