% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getAvaiableTemplates}
\alias{getAvaiableTemplates}
\title{Get avilable templates.}
\usage{
getAvaiableTemplates(list.out = FALSE)
}
\arguments{
\item{list.out}{Logical: If TRUE return a list of the full descriptions of the templates.}
}
\description{
Get avilable templates.
}
