% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioticUtilities.R
\name{getBioticLevels}
\alias{getBioticLevels}
\alias{getBioticKeys}
\title{Get the keys of a biotic XSD}
\usage{
getBioticLevels(xsd = "3.0")

getBioticKeys(xsd = "3.0")
}
\arguments{
\item{xsd}{A string identifying the XSD}
}
\value{
A vector of attributes or levels of the XSD.
}
\description{
This function gets all keys of a biotic XSD used by StoX.
}
\section{Functions}{
\itemize{
\item \code{getBioticKeys}: Get the levels of a biotic XSD
}}

