% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{getBootstrapData}
\alias{getBootstrapData}
\title{Get bootstrap data saved in a NetCDF4 file}
\usage{
getBootstrapData(
  filePath,
  nc,
  selection = list(),
  BootstrapIDStart = 1,
  BootstrapIDEnd = Inf,
  dropList1 = FALSE,
  close = TRUE
)
}
\arguments{
\item{filePath}{The path to the file.}

\item{nc}{A netCDF4 object, overiding the \code{filePath}.}

\item{selection}{Hierarchical list of names of the groups/variables. The last element must be a vector of the variables to return from the table specified by the other elements. E.g., list("ImputeSuperIndividuals", "SuperIndividualsData", c("Stratum", "IndividualAge", "Abundance")) will return a data.table of the three columns "Stratum", "IndividualAge" and "Abundance", added the BootstrapID specified in \code{BootstrapID}.}

\item{BootstrapIDStart, BootstrapIDEnd}{The start and end bootstrap IDs, i.e., the indices of the bootstrap replicates. The default returns all bootstrap replicates.}

\item{dropList1}{Logical: For use when a single table is requested. If FALSE (the default) return a list named by the \code{selection}. If TRUE return only the table.}

\item{close}{Logical: If TRUE (the default) close the netCDF4 file after reading the data.}
}
\description{
Get bootstrap data saved in a NetCDF4 file
}
