% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{getBootstrapNetCDF4Data}
\alias{getBootstrapNetCDF4Data}
\title{Get bootstrap data saved in a NetCDF4 file}
\usage{
getBootstrapNetCDF4Data(
  BootstrapNetCDF4Data,
  selection = list(),
  BootstrapIDStart = 1,
  BootstrapIDEnd = Inf
)
}
\arguments{
\item{BootstrapNetCDF4Data}{The \code{\link{BootstrapNetCDF4Data}} data output from \code{\link{BootstrapNetCDF4}}.}

\item{selection}{Hierarchical list of names of the groups/variables. The last element must be a vector of the variables to return from the table specified by the other elements. E.g., list("ImputeSuperIndividuals", "SuperIndividualsData", c("Stratum", "IndividualAge", "Abundance")) will return a data.table of the three columns "Stratum", "IndividualAge" and "Abundance", added the BootstrapID specified in \code{BootstrapID}.}

\item{BootstrapIDStart, BootstrapIDEnd}{The start and end bootstrap IDs, i.e., the indices of the bootstrap replicates. The default returns all bootstrap replicates.}
}
\description{
Get bootstrap data saved in a NetCDF4 file
}
