% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{getBootstrapNetCDF4ProcessNames}
\alias{getBootstrapNetCDF4ProcessNames}
\title{Get the process names of the processes stored in a BootstrapNetCDF4Data.}
\usage{
getBootstrapNetCDF4ProcessNames(BootstrapNetCDF4Data)
}
\arguments{
\item{BootstrapNetCDF4Data}{The \code{\link{BootstrapNetCDF4Data}} data output from \code{\link{BootstrapNetCDF4}}.}
}
\description{
For processes with function output as a list with the two elements Data and Resolution, only the Data are used in reports.
}
