% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{getBootstrapNetCDF4Variables}
\alias{getBootstrapNetCDF4Variables}
\title{Get the variable names of a BootstrapNetCDF4Data.}
\usage{
getBootstrapNetCDF4Variables(BootstrapNetCDF4Data, discardNrow = TRUE)
}
\arguments{
\item{BootstrapNetCDF4Data}{The \code{\link{BootstrapNetCDF4Data}} data output from \code{\link{BootstrapNetCDF4}}.}

\item{discardNrow}{Logical: If TRUE discard the nrow variable for each table.}
}
\description{
Get the variable names of a BootstrapNetCDF4Data.
}
