% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{getBootstrapOutputVariables}
\alias{getBootstrapOutputVariables}
\title{Get variables used in reports, specifically for use in the argument OutputVariables in Bootstrap()}
\usage{
getBootstrapOutputVariables(
  projectPath,
  parameters = c("GroupingVariables", "InformationVariables", "WeightingVariable",
    "TargetVariable")
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{parameters}{Character: A vector of strings naming the parameters to get from the ReportBootstrap functions.}
}
\description{
Get variables used in reports, specifically for use in the argument OutputVariables in Bootstrap()
}
