% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getEnabledAndFileOutputProcessNames}
\alias{getEnabledAndFileOutputProcessNames}
\title{Function to get all processes which are enabled and write output files:}
\usage{
getEnabledAndFileOutputProcessNames(projectPath)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}
}
\value{
A list with one vector of process names per model.
}
\description{
Function to get all processes which are enabled and write output files:
}
