% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getModelData}
\alias{getModelData}
\title{Get the data output data of a model.}
\usage{
getModelData(projectPath, modelName, startProcess = 1, endProcess = Inf)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{startProcess}{The process index of the first process to run.}

\item{endProcess}{The process index of the last process to run.}
}
\description{
This function is used by RstoxAPI::runModel().
}
