% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{getNonRstoxDependencies}
\alias{getNonRstoxDependencies}
\alias{downloadNonRstoxDependencies}
\title{Function to get the non Rstox dependencies of a package}
\usage{
getNonRstoxDependencies(
  packageName = c("RstoxFramework", "RstoxBase", "RstoxData"),
  dependencies = NA,
  Rstox.repos = NULL,
  nonRstox.repos = "https://cloud.r-project.org",
  sort = FALSE
)

downloadNonRstoxDependencies(
  packageName = c("RstoxFramework", "RstoxBase", "RstoxData"),
  destdir = NA,
  dependencies = NA,
  Rstox.repos = NULL,
  download.repos = "https://cloud.r-project.org",
  sort = FALSE,
  platform = NA,
  twoDigitRVersion = NA,
  ...
)
}
\arguments{
\item{packageName}{The packages to get dependencies from.}

\item{dependencies}{A character vector listing the dependencies to get, with possible values "Depends", "Imports", "LinkingTo", "Suggests", "Enhances". Default is NA, implying c("Depends", "Imports", "LinkingTo").}

\item{Rstox.repos, nonRstox.repos}{Either NULL to search for packages locally, or a repository passed to \code{\link[utils]{available.packages}}, where \code{Rstox.repos} is used to locate direct dependencies of the Rstox packages, and \code{nonRstox.reposs} is used to get the dependencies of the direct dependencies recursievly, defaulted to the CRAN repository.}

\item{sort}{Logical: If TRUE sort the dependent packages, defaulted to FALSE to enable installing the most basic dependencies first.}

\item{destdir}{The directory to download binaries to, defaulted to NA which implies tempdir().}
}
\description{
Function to get the non Rstox dependencies of a package
}
