% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{getOfficialRstoxPackageVersion}
\alias{getOfficialRstoxPackageVersion}
\title{Function to read the OfficialRstoxFrameworkVersions.txt file for the given StoX GUI version}
\usage{
getOfficialRstoxPackageVersion(
  StoXGUIVersion = NULL,
  officialRstoxPackageVersionsFile = system.file("versions",
    "OfficialRstoxFrameworkVersions.txt", package = "RstoxFramework"),
  packageName = c("RstoxFramework", "RstoxBase", "RstoxData"),
  include.optional = FALSE,
  toJSON = FALSE,
  list.out = FALSE
)
}
\arguments{
\item{StoXGUIVersion}{The version of the StoX GUI defining the combination of official Rstox package versions.}

\item{officialRstoxPackageVersionsFile}{The path to the file holding the link between StoX GUI version and Rstox package versions.}

\item{packageName}{The packages considered official Rstox pakcages, "RstoxFramework","RstoxBase" and "RstoxData".}

\item{include.optional}{Logical: If TRUE inclcude optional packages. Currently not used.}

\item{toJSON}{Logical: If TRUE output a JSON string.}
}
\description{
Function to read the OfficialRstoxFrameworkVersions.txt file for the given StoX GUI version
}
