% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes.R
\name{getParameterTableInfo}
\alias{getParameterTableInfo}
\title{GUI function: Function to get the info required for populating a parameter table builder in the GUI}
\usage{
getParameterTableInfo(projectPath, modelName, processID, format)
}
\description{
GUI function: Function to get the info required for populating a parameter table builder in the GUI
}
