% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes.R
\name{getParameterVectorInfo}
\alias{getParameterVectorInfo}
\title{GUI function: Function to get the info required for populating a parameter table builder in the GUI}
\usage{
getParameterVectorInfo(
  projectPath,
  modelName,
  processID,
  format,
  stopIfEmptyPossibleValues = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{format}{A character string naming the format to get info for.}
}
\description{
GUI function: Function to get the info required for populating a parameter table builder in the GUI
}
