% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getProcess}
\alias{getProcess}
\title{Get the properties of a StoX process}
\usage{
getProcess(projectPath, modelName, processName, only.valid = FALSE)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processName}{The name of the process.}

\item{only.valid}{Logical: If TRUE return only the valid arguments, which are those visible in the GUI.}
}
\description{
This function lists the process name, the function name, the process parameters, the function parameters, the function inputs and the process data of a StoX process.
}
