% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getProcessIDByOffset}
\alias{getProcessIDByOffset}
\title{Function to get process index from process ID}
\usage{
getProcessIDByOffset(projectPath, modelName, processID, offset = 0)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{offset}{An integer value to offset the process by, where -1 means the previous process and 1 means the next process.}
}
\description{
Function to get process index from process ID
}
