% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getProcessIDFromFunctionName}
\alias{getProcessIDFromFunctionName}
\title{Get process ID from function name}
\usage{
getProcessIDFromFunctionName(projectPath, modelName, functionName)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{functionName}{The name of the function used by the process.}
}
\description{
Get process ID from function name
}
