% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getProcessOutput}
\alias{getProcessOutput}
\title{Get output of a StoX process.}
\usage{
getProcessOutput(projectPath, modelName, processID, tableName = NULL,
  subFolder = NULL, flatten = FALSE, pretty = FALSE,
  linesPerPage = 1000L, pageindex = integer(0),
  search = character(0), drop = FALSE)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "Baseline", "Analysis" and "Report").}

\item{processID}{The ID of the process.}

\item{tableName}{The name of the table to extract from the process.}
}
\description{
Gets the output of a process that has been run.
}
