% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getProcessOutputFolder}
\alias{getProcessOutputFolder}
\title{Find the output folder of a process}
\usage{
getProcessOutputFolder(
  projectPath,
  modelName,
  processID,
  type = c("memory", "output", "text"),
  subFolder = NULL
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{type}{The type of output folder, one of "memory", to point to the memory output (files that only live while the project in open), or "output" or "text to point to the output folder holding files that continue living after the project is closed (written when Write output to file is checked in the GUI).}

\item{subFolder}{If the process returns subfolders (ReadBiotic and ReadAcoustic, where the subfolders represent files), specify the name of the folder with this parameter.}
}
\description{
Find the output folder of a process
}
