% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getRelevantClass}
\alias{getRelevantClass}
\title{Function to get the first class of the object x, except if the class "gg" is returned in which case "ggplot" is returned instead if present as one of the classes of the object:}
\usage{
getRelevantClass(x)
}
\arguments{
\item{x}{An R object}
}
\description{
Function to get the first class of the object x, except if the class "gg" is returned in which case "ggplot" is returned instead if present as one of the classes of the object:
}
