% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxFunctionAttributes.R
\name{getResamplableProcesses}
\alias{getResamplableProcesses}
\title{Utility function for processPropertyFormats. This is exported in order for processPropertyFormats to be albe to use it:}
\usage{
getResamplableProcesses(projectPath)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}
}
\description{
Utility function for processPropertyFormats. This is exported in order for processPropertyFormats to be albe to use it:
}
