% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxFunctionAttributes.R
\name{getResampleFunctions}
\alias{getResampleFunctions}
\title{Utility function for processPropertyFormats. This is exported in order for processPropertyFormats to be albe to use it:}
\usage{
getResampleFunctions()
}
\description{
Utility function for processPropertyFormats. This is exported in order for processPropertyFormats to be albe to use it:
}
