% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{hasFileOutput}
\alias{hasFileOutput}
\title{Does the process store output files?:}
\usage{
hasFileOutput(projectPath, modelName, processID, requireExists = TRUE)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{requireExists}{Logical: Should the existence of the output folder be checked.}
}
\description{
Does the process store output files?:
}
