% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Definitions.R
\name{initiateRstoxFramework}
\alias{initiateRstoxFramework}
\title{Definitions stored in the RstoxFramework environment}
\usage{
initiateRstoxFramework()
}
\value{
A list of definitions.
}
\description{
This function declares the RstoxFramework environment and writes vital definitions to it.
}
\seealso{
Use \code{\link{getRstoxFrameworkDefinitions}} to get the definitions.
}
