% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{installPackages}
\alias{installPackages}
\alias{getNonRstoxDependencies}
\alias{installOfficialRstoxPackagesWithDependencies}
\alias{getOfficialRstoxPackageVersion}
\title{Tools to download and install Rstox packages and dependen packages.}
\usage{
getNonRstoxDependencies(
  packageName = c("RstoxFramework", "RstoxBase", "RstoxData"),
  dependencyTypes = NA,
  Rstox.repos = NULL,
  nonRstox.repos = "https://cloud.r-project.org",
  sort = FALSE
)

installOfficialRstoxPackagesWithDependencies(
  StoXGUIVersion,
  officialRstoxPackageVersionsFile,
  destdir = NA,
  Rstox.repos = "https://stoxproject.github.io/repo",
  dependency.repos = "https://cloud.r-project.org",
  dependencyTypes = NA,
  lib = NULL,
  sort = FALSE,
  platform = NA,
  toJSON = FALSE,
  quiet = FALSE
)

getOfficialRstoxPackageVersion(
  StoXGUIVersion = NULL,
  officialRstoxPackageVersionsFile = system.file("versions",
    "OfficialRstoxFrameworkVersions.txt", package = "RstoxFramework"),
  packageName = c("RstoxFramework", "RstoxBase", "RstoxData"),
  toJSON = FALSE,
  list.out = FALSE
)
}
\arguments{
\item{packageName}{The packages considered official Rstox pakcages, "RstoxFramework","RstoxBase" and "RstoxData".}

\item{dependencyTypes}{A character vector listing the dependencies to get, with possible values "Depends", "Imports", "LinkingTo", "Suggests", "Enhances". Default is NA, implying c("Depends", "Imports", "LinkingTo").}

\item{Rstox.repos, nonRstox.repos}{Either NULL to search for packages locally, or a repository passed to \code{\link[utils]{available.packages}}, where \code{Rstox.repos} is used to locate direct dependencies of the Rstox packages, and \code{nonRstox.repos} is used to get the dependencies of the direct dependencies recursively, defaulted to the CRAN repository.}

\item{sort}{Logical: If TRUE sort the dependent packages, defaulted to FALSE to enable installing the most basic dependencies first.}

\item{StoXGUIVersion}{The version of the StoX GUI defining the combination of official Rstox package versions.}

\item{officialRstoxPackageVersionsFile}{The path to the file holding the link between StoX GUI version and Rstox package versions.}

\item{destdir}{The directory to download binaries to, defaulted to NA which implies tempdir().}

\item{dependency.repos}{The repository to download from.}

\item{lib}{
    character vector giving the library directories where to
    install the packages.  Recycled as needed.  If missing, defaults to
    the first element of \code{\link{.libPaths}()}.
  }

\item{platform}{The platform to download binaries for, defaulted to the current platform, but with the option to download for other platforms (possible values are "windows" and "macosx").}

\item{toJSON}{Logical: If TRUE output a JSON string.}

\item{quiet}{
    logical: if true, reduce the amount of output.  This is \emph{not}
    passed to \code{\link[utils]{available.packages}()} in case that is called, on
    purpose.
  }

\item{list.out}{Logical: If TRUE wrap the output of \code{getOfficialRstoxPackageVersion} in a list with packageName and version.}
}
\description{
\code{getNonRstoxDependencies} gets all package dependencies ("Depends", "Imports" and "LinkingTo") of the packages given by \code{packageName}, excluding any Rstox-packages. \cr \cr
\code{installOfficialRstoxPackagesWithDependencies} installs the corrrect versions of Rstox-packages, and other packages that these packages depend on. This function is used by the StoX GUI. \cr \cr
\code{getOfficialRstoxPackageVersion} reads the OfficialRstoxFrameworkVersions.txt file for the given StoX GUI version \cr \cr
}
