% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{isCertifiedRstoxFramework}
\alias{isCertifiedRstoxFramework}
\title{Test for whether the installed Rstox packages are certified.}
\usage{
isCertifiedRstoxFramework(
  StoXVersion,
  officialRstoxPackageVersionsFile,
  optionalDependencies = FALSE
)
}
\arguments{
\item{StoXVersion}{The version of the StoX GUI defining the combination of official Rstox package versions.}

\item{officialRstoxPackageVersionsFile}{The path to the file holding the link between StoX GUI version and Rstox package versions. If missing, the file on the RstoxFramework master on GitHub is used ("https://raw.githubusercontent.com/StoXProject/RstoxFramework/master/inst/versions/OfficialRstoxFrameworkVersions.txt").}

\item{optionalDependencies}{Logical: If TRUE include also the column OptionalDependencies from the OfficialRstoxFrameworkVersions.txt file, which holds Rstox packages used in the Sugggests field of the DESCRIPTION file.}
}
\description{
This function is used by the StoX GUI to set colour to the RstoxFramework logo.
}
