% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{isOfficialStoXVersion}
\alias{isOfficialStoXVersion}
\title{Test for whether the given StoX version is an official version}
\usage{
isOfficialStoXVersion(StoXVersion, officialRstoxPackageVersionsFile)
}
\arguments{
\item{StoXVersion}{The version of the StoX GUI defining the combination of official Rstox package versions.}

\item{officialRstoxPackageVersionsFile}{The path to the file holding the link between StoX GUI version and Rstox package versions. If missing, the file on the RstoxFramework master on GitHub is used ("https://raw.githubusercontent.com/StoXProject/RstoxFramework/master/inst/versions/OfficialRstoxFrameworkVersions.txt").}
}
\description{
Test for whether the given StoX version is an official version
}
