% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectIO.R
\name{isStox27project}
\alias{isStox27project}
\title{valid stox 2.7 project}
\usage{
isStox27project(stox27project)
}
\arguments{
\item{stox27project}{object to check for validity}
}
\value{
logical true if 'stox27project' is a valid ~\code{\link[RstoxFramework]{stox27project}}
}
\description{
checks if an object is a valid ~\code{\link[RstoxFramework]{stox27project}}
}
