% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{json2expression}
\alias{json2expression}
\title{Function to convert from json to R expression:}
\usage{
json2expression(json)
}
\arguments{
\item{json}{A JSON string}
}
\description{
Function to convert from json to R expression:
}
