% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{list2expression}
\alias{list2expression}
\title{Function to convert from R list to R expression:}
\usage{
list2expression(l, parentHasSiblings = FALSE)
}
\arguments{
\item{l}{An R list.}

\item{parentHasSiblings}{Logical: If TRUE there are more than one element in the top level of the list.}
}
\description{
Function to convert from R list to R expression:
}
