% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ListFilesInFolder}
\alias{ListFilesInFolder}
\title{List files in a folder}
\usage{
ListFilesInFolder(FolderName = character())
}
\arguments{
\item{FolderName}{The path to the folder}
}
\description{
List files in a folder
}
