% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ModelData}
\alias{ModelData}
\title{Model data generated by RstoxFramework}
\arguments{
\item{BootstrapData}{The \code{\link{BootstrapData}} data.}

\item{ReportBootstrapData}{The \code{\link{ReportBootstrapData}} data.}

\item{PlotReportBootstrapData}{The \code{\link{PlotReportBootstrapData}} data.}

\item{StratumAreaData}{The \code{\link[RstoxBase]{StratumAreaData}} data (defined in RstoxBase).}

\item{NASCData}{The \code{\link[RstoxBase]{NASCData}} data (defined in RstoxBase).}

\item{SumNASCData}{The \code{\link[RstoxBase]{SumNASCData}} data (defined in RstoxBase).}

\item{MeanNASCData}{The \code{\link[RstoxBase]{MeanNASCData}} data (defined in RstoxBase).}

\item{LengthDistributionData}{The \code{\link[RstoxBase]{LengthDistributionData}} data (defined in RstoxBase).}

\item{SumLengthDistributionData}{The \code{\link[RstoxBase]{SumLengthDistributionData}} data (defined in RstoxBase).}

\item{MeanLengthDistributionData}{The \code{\link[RstoxBase]{MeanLengthDistributionData}} data (defined in RstoxBase).}

\item{AssignmentLengthDistributionData}{The \code{\link[RstoxBase]{AssignmentLengthDistributionData}} data (defined in RstoxBase).}

\item{DensityData}{The \code{\link[RstoxBase]{DensityData}} data (defined in RstoxBase).}

\item{MeanDensityData}{The \code{\link[RstoxBase]{MeanDensityData}} data (defined in RstoxBase).}

\item{SpeciesCategoryCatchData}{The \code{\link[RstoxBase]{SpeciesCategoryCatchData}} data (defined in RstoxBase).}

\item{SumSpeciesCategoryCatchData}{The \code{\link[RstoxBase]{SumSpeciesCategoryCatchData}} data (defined in RstoxBase).}

\item{MeanSpeciesCategoryCatchData}{The \code{\link[RstoxBase]{MeanSpeciesCategoryCatchData}} data (defined in RstoxBase).}

\item{QuantityData}{The \code{\link[RstoxBase]{QuantityData}} data (defined in RstoxBase).}

\item{IndividualsData}{The \code{\link[RstoxBase]{IndividualsData}} data (defined in RstoxBase).}

\item{SuperIndividualsData}{The \code{\link[RstoxBase]{SuperIndividualsData}} data (defined in RstoxBase).}

\item{ReportSuperIndividualsData}{The \code{\link[RstoxBase]{ReportSuperIndividualsData}} data (defined in RstoxBase).}

\item{PlotAcousticTrawlSurveyData}{The \code{\link[RstoxBase]{PlotAcousticTrawlSurveyData}} data (defined in RstoxBase).}

\item{BioticData}{\code{\link[RstoxData]{BioticData}} (defined in RstoxData).}

\item{AcousticData}{\code{\link[RstoxData]{AcousticData}} (defined in RstoxData).}

\item{LandingData}{\code{\link[RstoxData]{LandingData}} (defined in RstoxData).}

\item{StoxBioticData}{\code{\link[RstoxData]{StoxBioticData}} (defined in RstoxData).}

\item{StoxAcousticData}{\code{\link[RstoxData]{StoxAcousticData}} (defined in RstoxData).}

\item{StoxLandingData}{\code{\link[RstoxData]{StoxLandingData}} (defined in RstoxData).}

\item{MergeStoxBioticData}{\code{\link[RstoxData]{MergeStoxBioticData}} (defined in RstoxData).}

\item{MergeStoxAcousticData}{\code{\link[RstoxData]{MergeStoxAcousticData}} (defined in RstoxData).}

\item{ICESBioticData}{\code{\link[RstoxData]{ICESBioticData}} (defined in RstoxData).}

\item{ICESAcousticData}{\code{\link[RstoxData]{ICESAcousticData}} (defined in RstoxData).}

\item{ICESDatrasData}{\code{\link[RstoxData]{ICESDatrasData}} (defined in RstoxData).}

\item{ICESDatsuscData}{\code{\link[RstoxData]{ICESDatsuscData}} (defined in RstoxData).}

\item{WriteICESBioticData}{\code{\link[RstoxData]{WriteICESBioticData}} (defined in RstoxData).}

\item{WriteICESAcousticData}{\code{\link[RstoxData]{WriteICESAcousticData}} (defined in RstoxData).}

\item{WriteICESDatrasData}{\code{\link[RstoxData]{WriteICESDatrasData}} (defined in RstoxData).}

\item{WriteICESDatsuscData}{\code{\link[RstoxData]{WriteICESDatsuscData}} (defined in RstoxData).}
}
\description{
The model data of a StoX model are the data generated during the model run based on input data and user settings and resources given in the project description (project.json file). Model data are transient and only exists from a process has been run until the project is closed.
}
\seealso{
\code{\link{ProcessData}} for process data types and \code{\link{DataTypes}} for all data types produced by \code{\link{RstoxFramework}}. See also \code{\link[RstoxData]{ModelData}} in RstoxData and \code{\link[RstoxBase]{ModelData}} in RstoxBase.
}
