% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{modifyProcess}
\alias{modifyProcess}
\title{Modify a process}
\usage{
modifyProcess(
  projectPath,
  modelName,
  processName,
  newValues,
  archive = TRUE,
  add.defaults = FALSE,
  purge.processData = FALSE,
  strict = TRUE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}

\item{processName}{The name of the process(es).}

\item{newValues}{A list named by the elements to modify (see getRstoxFrameworkDefinitions("processProperties") for possible elements), holding the values to modify to (e.g., list(functionParameter = list(DefinitionMethod = "Stratum"))).}

\item{archive}{Logical: Should the project memory state be archived using \code{archiveProject}? This stores one step in the history.}

\item{add.defaults}{Logical: If TRUE defaults of a function are added when setting the function of the process.}

\item{purge.processData}{Logical: If TRUE replace process data entirely.}

\item{strict}{Logical: If FALSE a proposed process nname is changed to the default new process name, whereas TRUE throws an error.}
}
\description{
Modify a process
}
