% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{modifyProject}
\alias{modifyProject}
\title{Modify a project}
\usage{
modifyProject(
  projectPath,
  modelNames = getRstoxFrameworkDefinitions("stoxModelNames"),
  newValues,
  add.defaults = FALSE,
  purge.processData = FALSE,
  strict = TRUE,
  update.functionInputs = TRUE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelNames}{The name of the models (possible values are "baseline", "analysis" and "report").}

\item{newValues}{A list named by the elements to modify (see getRstoxFrameworkDefinitions("processProperties") for possible elements), holding the values to modify to (e.g., list(functionParameter = list(DefinitionMethod = "Stratum"))).}

\item{add.defaults}{Logical: If TRUE defaults of a function are added when setting the function of the process.}

\item{purge.processData}{Logical: If TRUE replace process data entirely.}

\item{strict}{Logical: If FALSE a proposed process name is changed to the default new process name, whereas TRUE throws an error.}

\item{update.functionInputs}{Logical: If TRUE update the process name in the function inputs to other processes.}
}
\description{
Modify a project
}
