% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{modifyProjects}
\alias{modifyProjects}
\title{Modify multiple projects}
\usage{
modifyProjects(
  projectPaths,
  modelNames = getRstoxFrameworkDefinitions("stoxModelNames"),
  newValuesList,
  newValuesListByProject,
  add.defaults = FALSE,
  purge.processData = FALSE,
  strict = TRUE,
  update.functionInputs = TRUE
)
}
\arguments{
\item{projectPaths}{A vector of paths to the StoX projects.}

\item{modelNames}{The name of the models (possible values are "baseline", "analysis" and "report").}

\item{newValuesList}{A list of \code{newValues} named by the processes to modify. See also \code{\link{modifyProcess}}.}

\item{newValuesListByProject}{A list of \code{newValuesList}, one for each project. If the same value should be used on all projects, use \code{newValuesList} instead.}

\item{add.defaults}{Logical: If TRUE defaults of a function are added when setting the function of the process.}

\item{purge.processData}{Logical: If TRUE replace process data entirely.}

\item{strict}{Logical: If FALSE a proposed process name is changed to the default new process name, whereas TRUE throws an error.}

\item{update.functionInputs}{Logical: If TRUE update the process name in the function inputs to other processes.}
}
\description{
Modify multiple projects
}
