% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{parseParameter}
\alias{parseParameter}
\title{Function to parse a parameter coming from the GUI}
\usage{
parseParameter(parameter, simplifyVector = TRUE)
}
\arguments{
\item{parameter}{}

\item{simplifyVector}{}
}
\description{
Function to parse a parameter coming from the GUI
}
