% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{PlotReportBootstrap}
\alias{PlotReportBootstrap}
\title{Plot ReportBootstrapData}
\usage{
PlotReportBootstrap(
  ReportBootstrapData,
  PlotType = c("ErrorBarPlot"),
  PlottingVariable = character(),
  PlottingVariableLower = character(),
  PlottingVariableUpper = character(),
  AddCVToPlot = FALSE,
  CVVariable = character(),
  GroupingVariables = character(),
  SubPlots = character(),
  UseDefaultLabelSettings = TRUE,
  Title = character(),
  AxisTitleSize = numeric(),
  AxisTickSize = numeric(),
  LegendTitleSize = numeric(),
  LegendTextSize = numeric(),
  UseDefaultFileSettings = TRUE,
  Format = character(),
  Width = numeric(),
  Height = numeric(),
  DotsPerInch = numeric()
)
}
\arguments{
\item{ReportBootstrapData}{The \code{\link{ReportBootstrapData}} data.}

\item{PlotType}{The type of plot to produce. Currently only "ErrorBarPlot" is implemented.}

\item{PlottingVariable}{The variable to plot the points for. Normally this is the variable giving the mean.}

\item{PlottingVariableLower}{The variable to use for the lower end of the error bars.}

\item{PlottingVariableUpper}{The variable to use for the upper end of the error bars.}

\item{AddCVToPlot}{Logical: If TRUE add a dotted line with points for the coefficient of variation (standard deviation divided by mean) CV. Setting this to TRUE requires to specify the \code{CVVariable}.}

\item{CVVariable}{The name of the variable holding the CV. Must be given if \code{AddCVToPlot} is TRUE.}

\item{GroupingVariables}{A vector of length 1 or 2 giving the names of the variables to plot along the x axis, where the second is coded with colors in the plot.}

\item{SubPlots}{A vector of the names of the plots to produce. To get a list of possible plot names use getSubPlotNames_PlotReportBootstrap(ReportBootstrapData, GroupingVariables). The subplot names cannot contain slashes, which are replaced by underscore.}

\item{UseDefaultLabelSettings}{Logical: If TRUE (default) use the default text settings of the plot, specifically Title, AxisTitleSize, AxisTickSize, LegendTitleSize and LegendTextSize. Setting this to FALSE will show all text options of the plotting function in the StoX GUI.}

\item{Title}{Character: The title of the plot, defaulted to no title.}

\item{AxisTitleSize}{Numeric: The size of the axis titles, defaulted to 20.}

\item{AxisTickSize}{Numeric: The size of the axis tick marks, defaulted to 20.}

\item{LegendTitleSize}{Numeric: The size of the legend title, defaulted to 20.}

\item{LegendTextSize}{Numeric: The size of the legend text, defaulted to 20.}

\item{UseDefaultFileSettings}{Logical: If TRUE (default) use the default file settings of the output file of the plot, specifically Format, Width, Height and DotsPerInch. Setting this to FALSE will show all file options of the plotting function in the StoX GUI.}

\item{Format}{Character: The format of the output file from plotting functions, defaulted to "png". Other options are defined by \code{\link[ggplot2]{ggsave}}, including "tiff", "jpeg"  and "pdf".}

\item{Width}{Numeric: The width of the plot in centimeters, defaulted to 17, as per the instructions to authors for submission to the ICES Journal og Marine Science.}

\item{Height}{Numeric: The height of the plot in centimeters, defaulted to 17.}

\item{DotsPerInch}{Numeric: The number of dots per inch (DPI), defaulted to 1200.}
}
\value{
A \code{\link{PlotReportBootstrapData}} object.
}
\description{
Plot ReportBootstrapData
}
