% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{PlotReportBootstrapData}
\alias{PlotReportBootstrapData}
\title{Plot of reported Bootstrap data}
\description{
The PlotReportBootstrapData model data is a ggplot object or a list of ggplot objects generated from \code{\link{BootstrapData}}. The plots show e.g. error bars with lower, mid and upper value of Abundance per IndividualAge.
}
\seealso{
This data type is produced by \code{\link{PlotReportBootstrap}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}, \code{\link{RstoxBase}} and \code{\link{RstoxFramework}}.
}
