% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{prependProcess}
\alias{prependProcess}
\title{Prepend (append before) to a process by adding another process of the same function}
\usage{
prependProcess(
  projectPath,
  modelName,
  processName,
  prependProcessName = paste(processName, "prepended", sep = "_"),
  values = NULL,
  returnProcessTable = TRUE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processName}{The name of the process.}

\item{prependProcessName}{The name of process to prepend.}

\item{values}{A list of zero or more of the following process arguments: functionParameters, processData, processParameters. The arguments functionInputs, functionName and processName are set as equal to the process given by \code{processName}.}

\item{returnProcessTable}{Logical: If TRUE return the process table of the specified model (much used in functions used by the GUI). If NULL returns the process table of all models.}
}
\description{
This method is only available for functions that has the same data type as input and output.
}
