% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ProcessData}
\alias{ProcessData}
\title{Process data used by RstoxFramework}
\arguments{
\item{StratumPolygon}{The \code{\link[RstoxBase]{StratumPolygon}} process data (defined in RstoxBase).}

\item{Survey}{The \code{\link[RstoxBase]{Survey}} process data (defined in RstoxBase).}

\item{AcousticLayer}{The \code{\link[RstoxBase]{AcousticLayer}} process data (defined in RstoxBase).}

\item{AcousticPSU}{The \code{\link[RstoxBase]{AcousticPSU}} process data (defined in RstoxBase).}

\item{BioticLayer}{The \code{\link[RstoxBase]{BioticLayer}} process data (defined in RstoxBase).}

\item{BioticPSU}{The \code{\link[RstoxBase]{BioticPSU}} process data (defined in RstoxBase).}

\item{BioticAssignment}{The \code{\link[RstoxBase]{BioticAssignment}} process data (defined in RstoxBase).}

\item{AcousticTargetStrength}{The \code{\link[RstoxBase]{AcousticTargetStrength}} process data (defined in RstoxBase).}

\item{Regression}{The \code{\link[RstoxBase]{Regression}} process data (defined in RstoxBase).}

\item{Translation}{\code{\link[RstoxData]{Translation}} (defined in RstoxData).}
}
\description{
There are currently no process data in RstoxFramework, but in the dependent packages RstoxData and RstoxBase.
}
\seealso{
\code{\link{ModelData}} for model data types and \code{\link{DataTypes}} for all data types produced by \code{\link[RstoxFramework]{RstoxFramework}}.
}
