% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxFunctionAttributes.R
\name{processPropertyFormats}
\alias{processPropertyFormats}
\title{Process property formats for RstoxFramework}
\usage{
processPropertyFormats
}
\description{
Process property formats for RstoxFramework
}
