% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{progressOfProcesses}
\alias{progressOfProcesses}
\title{Get process progress}
\usage{
progressOfProcesses(projectPath, modelName, percent = FALSE)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}

\item{percent}{Logical: If TRUE return the progress in percent, otherwise in [0,1].}
}
\description{
Get process progress
}
