% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{purgeOutput}
\alias{purgeOutput}
\title{Delete the contents of the output folder of a model.}
\usage{
purgeOutput(projectPath, modelName)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}
}
\description{
This function is run by \code{\link{runModel}} to clear off any files present in the output folder of a model. It should also be used by GUIs when running a model.
}
