% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReadAcousticXML}
\alias{ReadAcousticXML}
\title{Read acoustic XML files}
\usage{
ReadAcousticXML(FileNames = NA, acousticFormat = "25")
}
\arguments{
\item{FileNames}{A vector of paths to XML files.}

\item{acousticFormat}{The format of the acoustic data, referring to the List User File (LUF) naming used by the Large Scale Survey System (LSSS) ************** refer to ICES instead!!!!!!!!!!!!!!}
}
\value{
A list of data.tables in the specified format.
}
\description{
This function reads multiple acoustic XML files.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[RNMDAPI]{readNMDxmlFile}}.
}
