% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReadBiotic}
\alias{ReadBiotic}
\title{Some title}
\usage{
ReadBiotic(FileNames)
}
\arguments{
\item{FileNames}{One or more file paths given either as (1) a vector, (2) a JSON string representing a vector, or (3) a comma or semicolon separated string of file paths.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
Some description
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
