% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReadLandingXML}
\alias{ReadLandingXML}
\title{Read landing XML files}
\usage{
ReadLandingXML(FileNames, landingFormat = NULL)
}
\arguments{
\item{FileNames}{A vector of paths to XML files.}

\item{landingFormat}{The format of the landing data.}
}
\value{
A list of data.tables in the specified format.
}
\description{
This function reads multiple landing XML files.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[RNMDAPI]{readNMDxmlFile}}.
}
