% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{readModelData}
\alias{readModelData}
\title{Read the output files of a project}
\usage{
readModelData(
  projectPath,
  modelName = NULL,
  processName = NULL,
  verifyFiles = FALSE,
  unlist.models = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}

\item{processName}{The name of the process(es).}

\item{verifyFiles}{Logical: If TRUE verify that the files are from processes that exist in the project.}

\item{unlist.models}{Logical: If TRUE unlist the top level so that all processes are in one list.}
}
\value{
A list of model output.
}
\description{
This function reads all or some of the output files of a project, indicated by model and proccess names.
}
