% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{readOfficialRstoxPackageVersionsFile}
\alias{readOfficialRstoxPackageVersionsFile}
\title{Get certified Rstox package versions}
\usage{
readOfficialRstoxPackageVersionsFile(
  officialRstoxPackageVersionsFile,
  optionalDependencies = FALSE,
  toTable = FALSE
)
}
\arguments{
\item{officialRstoxPackageVersionsFile}{The path to the file holding the link between StoX GUI version and Rstox package versions. If missing, the file on the RstoxFramework master on GitHub is used ("https://raw.githubusercontent.com/StoXProject/RstoxFramework/master/inst/versions/OfficialRstoxFrameworkVersions.txt").}

\item{optionalDependencies}{Logical: If TRUE include also the column OptionalDependencies from the OfficialRstoxFrameworkVersions.txt file, which holds Rstox packages used in the Suggests field of the DESCRIPTION file.}

\item{toTable}{Logical: If TRUE output a table.}
}
\description{
This function reads the OfficialRstoxFrameworkVersions.txt file for the given StoX GUI version.
}
