% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectIO.R
\name{readProject}
\alias{readProject}
\title{Read Stox project from project xml}
\usage{
readProject(projectxml)
}
\arguments{
\item{projectxml}{xml filename}
}
\value{
Nested list representation of project
}
\description{
Read Stox project from project xml
}
