% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectIO.R
\name{readStox27Project}
\alias{readStox27Project}
\title{Read Stox project from project xml}
\usage{
readStox27Project(projectxml, strict = T)
}
\arguments{
\item{projectxml}{xml filename}

\item{strict}{logical, whether errors should be raised for unkown elements and attributes or missing required attributes, if False warnings will be issued.}
}
\value{
Nested list representation of project, formatted as: \code{\link[RstoxFramework]{stox27project}}
}
\description{
Read Stox project from project xml
}
