% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataUtilities.R
\name{readXmlFiles}
\alias{readXmlFiles}
\title{Read XML files}
\usage{
readXmlFiles(FileNames, format = "biotic_3.0")
}
\arguments{
\item{FileNames}{A vector of paths to XML files.}

\item{format}{A string specifying the format to read the files in (converted if possible).}
}
\value{
A list of data.tables in the specified format.
}
\description{
This function reads multiple XML files and converts to the specified format.
}
\details{
The function \code{\link[RstoxData]{readXmlFile}} detects the format and reads a biotic, acoustic or landing XML file, and is used to read multiple files of the same type. The output is converted to the specified format, and data from all files are rbinded (stacked vertically).
}
\examples{
x <- 1

}
\seealso{
\code{\link[RstoxData]{readXmlFile}}.
}
