% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{rearrangeProcesses}
\alias{rearrangeProcesses}
\title{Select a number of processes and move these to a different location in a model.}
\usage{
rearrangeProcesses(projectPath, modelName, processID, afterProcessID = NULL)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{afterProcessID}{The process ID after which to more the specified processes to.}
}
\description{
Select a number of processes and move these to a different location in a model.
}
