% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{removeProcess}
\alias{removeProcess}
\title{Remove a StoX process}
\usage{
removeProcess(projectPath, modelName, processID)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}
}
\description{
Remove a StoX process
}
