% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{removeProcesses}
\alias{removeProcesses}
\title{Remove a StoX process}
\usage{
removeProcesses(projectPath, processNames)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{processNames}{A vector of names of the processes.}
}
\description{
Remove a StoX process
}
