% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{ReportBootstrapNetCDF4Data}
\alias{ReportBootstrapNetCDF4Data}
\title{Reported BootstrapNetCDF4 data}
\description{
The ReportBootstrapNetCDF4Data model data is a report of the \code{\link{BootstrapNetCDF4Data}} such as cv of Biomass per Stratum and SpeciesCategory, given as a table or list of tables.
}
\seealso{
This data type is produced by \code{\link{ReportBootstrapNetCDF4}}. See \code{\link{DataTypes}} for a list of all StoX data types produced by \code{\link{RstoxData}}, \code{\link{RstoxBase}} and \code{\link{RstoxFramework}}.
}
