% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{Resample_BioticAssignment_ByStratum}
\alias{Resample_BioticAssignment_ByStratum}
\title{Resamples biotic PSUs}
\usage{
Resample_BioticAssignment_ByStratum(BioticAssignment, Seed)
}
\arguments{
\item{BioticAssignment}{The \code{\link[RstoxBase]{BioticAssignment}} process data.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
This function resamples biotic PSUs with replacement within each Stratum, changing the MeanLengthDistributionWeight.
}
\seealso{
\code{\link[RstoxBase]{ModelData}} for model data types and \code{\link[RstoxBase]{DataTypes}} for all data types produced by \code{\link[RstoxBase]{RstoxBase}}.
}
