% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{Resample_MeanLengthDistributionData}
\alias{Resample_MeanLengthDistributionData}
\title{Resamples biotic PSUs within Stratum in MeanLengthDistributionData}
\usage{
Resample_MeanLengthDistributionData(MeanLengthDistributionData, Seed)
}
\arguments{
\item{MeanLengthDistributionData}{The \code{\link[RstoxBase]{MeanLengthDistributionData}} data.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
This function resamples biotic PSUs with replacement within each Stratum, changing the MeanLengthDistributionWeight.
}
