% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{Resample_MeanNASCData}
\alias{Resample_MeanNASCData}
\title{Resamples acoustic PSUs}
\usage{
Resample_MeanNASCData(MeanNASCData, Seed)
}
\arguments{
\item{MeanNASCData}{The \code{\link[RstoxBase]{MeanNASCData}} data.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
This function resamples acoustic PSUs with replacement within each Stratum, changing the MeanNASC
}
