% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{Resample_MeanSpeciesCategoryCatchData}
\alias{Resample_MeanSpeciesCategoryCatchData}
\title{Resamples biotic PSUs within Stratum in MeanSpeciesCategoryCatchData}
\usage{
Resample_MeanSpeciesCategoryCatchData(MeanSpeciesCategoryCatchData, Seed)
}
\arguments{
\item{MeanSpeciesCategoryCatchData}{The \code{\link[RstoxBase]{MeanSpeciesCategoryCatchData}} data.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
This function resamples biotic PSUs with replacement within each Stratum, changing the MeanSpeciesCategoryCatchData
}
