% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{Resample_PreySpeciesCategoryCatchData_Hierarchical_UsingScaling}
\alias{Resample_PreySpeciesCategoryCatchData_Hierarchical_UsingScaling}
\title{Resamples biotic PSUs within Stratum in MeanSpeciesCategoryCatchData}
\usage{
Resample_PreySpeciesCategoryCatchData_Hierarchical_UsingScaling(
  PreySpeciesCategoryCatchData,
  Seed
)
}
\arguments{
\item{PreySpeciesCategoryCatchData}{The \code{\link[RstoxBase]{PreySpeciesCategoryCatchData}} data.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
This function resamples biotic PSUs with replacement within each Stratum, changing the MeanSpeciesCategoryCatchData
}
