% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleBioticAssignment}
\alias{ResampleBioticAssignment}
\title{Resamples biotic PSUs}
\usage{
ResampleBioticAssignment(BioticAssignment, Seed)
}
\description{
This function resamples biotic PSUs with replacement within each Stratum, changing the MeanLengthDistributionWeight.
}
