% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleBioticPSUs}
\alias{ResampleBioticPSUs}
\title{Resamples biotic PSUs}
\usage{
ResampleBioticPSUs(MeanLengthDistributionData, Seed)
}
\arguments{
\item{MeanLengthDistributionData}{The \code{\link[RstoxBase]{MeanLengthDistributionData}} data.}

\item{Seed}{The seed, given as a sinigle initeger.}
}
\description{
This function resamples biotic PSUs with replacement within each Stratum, changing the MeanLengthDistributionWeight.
}
