% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resampleDataBy}
\alias{resampleDataBy}
\title{Resamples StoX data}
\usage{
resampleDataBy(data, seed, varToScale, varToResample, resampleBy)
}
\description{
This function resamples PSUs with replacement by altering the input data
}
