% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleEDSUs}
\alias{ResampleEDSUs}
\title{Resamples EDSUs}
\usage{
ResampleEDSUs(NASCData, Seed, ResampleBy = "Stratum")
}
\arguments{
\item{NASCData}{The \code{\link[RstoxBase]{NASCData}} to resample EDSUs in.}
}
\description{
This function resamples Hauls with replacement by altering the LengthDistributionWeight.
}
