% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resampleHauls}
\alias{resampleHauls}
\title{Resamples Hauls.}
\usage{
resampleHauls(LengthDistributionData, seed, by = "Stratum")
}
\arguments{
\item{LengthDistributionData}{The \code{\link[RstoxBase]{LengthDistributionData}} to resample Hauls in.}
}
\description{
This function resamples Hauls with replacement by altering the LengthDistributionWeight.
}
