% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resampleHaulsByStratum}
\alias{resampleHaulsByStratum}
\title{Resamples Hauls.}
\usage{
resampleHaulsByStratum(data, seed)
}
\arguments{
\item{data}{The \code{\link[RstoxBase]{LengthDistributionData}} to resample Hauls in.}
}
\description{
This function resamples Hauls with replacement by altering the LengthDistributionWeight.
}
