% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleMeanLengthDistributionData}
\alias{ResampleMeanLengthDistributionData}
\title{Resamples biotic PSUs within Stratum in MeanLengthDistributionData}
\usage{
ResampleMeanLengthDistributionData(MeanLengthDistributionData, Seed)
}
\arguments{
\item{MeanLengthDistributionData}{The \code{\link[RstoxBase]{MeanLengthDistributionData}} data.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
This function resamples, with replacement, the BioticPSUs within Stratum in the \code{\link[RstoxBase]{MeanLengthDistributionData}}, where Stratum is the stratum associated to the PSU, and not necessarily the actual stratum polygon. The column WeightedNumber is scaled by the number of occurrences of each BioticPSUs from the resampling.
}
