% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleMeanSpeciesCategoryCatchData}
\alias{ResampleMeanSpeciesCategoryCatchData}
\title{Resamples biotic PSUs within Stratum in MeanSpeciesCategoryCatchData}
\usage{
ResampleMeanSpeciesCategoryCatchData(MeanSpeciesCategoryCatchData, Seed)
}
\arguments{
\item{MeanSpeciesCategoryCatchData}{The \code{\link[RstoxBase]{MeanSpeciesCategoryCatchData}} data.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
This function resamples, with replacement, the BioticPSUs within Stratum in the \code{\link{MeanSpeciesCategoryCatchData}}, where Stratum is the stratum associated to the PSU, and not necessarily the actual stratum polygon. The columns TotalCatchWeight and TotalCatchNumber are scaled by the number of occurrences of each BioticPSUs from the resampling.
}
