% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{resetModel}
\alias{resetModel}
\title{Reset a StoX model.}
\usage{
resetModel(
  projectPath,
  modelName,
  processID = NULL,
  processDirty = FALSE,
  shift = 0,
  returnProcessTable = FALSE,
  purgeOutputFiles = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{processDirty}{Logical: Indicates whether the model has been modified when reseting.}

\item{shift}{The position relative to the current memory status to un/redo to.}
}
\description{
Reset a StoX model.
}
