% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RstoxFramework-package.R
\docType{package}
\name{RstoxFramework}
\alias{RstoxFramework}
\alias{RstoxFramework-package}
\title{The Engine of StoX}
\description{
The framwork of StoX >= 3.0.
}
\details{
The RstoxFramework package is the engine of the stock assesment utility StoX, which is an open source application fo acoustic-trawl and swept-area survey estimation.

The package creates an evironment containing the StoX project(s) in separate environments. Each StoX project consists of a list of data tables holding e.g. biotic and acoustic data, filtered versions of the data, strata system, definitios of primary sampling units, accompanied by a list of specifications of the StoX processes comprising the StoX project. A StoX process is an R function taking as input the project name, the input data and parameters used by the function.

The package replaces the old Java library in StoX versions prior to StoX 4.0.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StoXProject/RstoxFramework}
  \item Report bugs at \url{https://github.com/StoXProject/RstoxFramework/issues}
}

}
\author{
\strong{Maintainer}: Arne Johannes Holmin \email{arnejh@hi.no}

Authors:
\itemize{
  \item Ibrahim Umar
  \item Edvin Fuglebakk
  \item Aasmund Skaalevik
  \item Esmael Musema Hassen
  \item Sindre Vatnehol
  \item Espen Johnsen
  \item Atle Totland
  \item Mikko Juhani Vihtakari
}

Other contributors:
\itemize{
  \item Norwegian Institute of Marine Research [copyright holder, funder]
}

}
