% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{runFunction}
\alias{runFunction}
\alias{runFunction.JSON}
\title{General functions to run a function of an Rstox package, modeled by do.call().}
\usage{
runFunction(
  what,
  args,
  package = "RstoxFramework",
  removeCall = TRUE,
  onlyStoxMessages = TRUE,
  onlyStoxWarnings = TRUE,
  onlyStoxErrors = FALSE,
  maxLength.Message = 2000,
  maxLength.Warning = 2000,
  maxLength.Error = 2000,
  nwarnings = 10000,
  collapseMessages = TRUE,
  collapseWarnings = TRUE,
  collapseErrors = TRUE,
  messageLogFile = NULL,
  warningLogFile = NULL,
  errorLogFile = NULL,
  messageAppend = FALSE,
  warningAppend = FALSE,
  errorAppend = FALSE,
  uniquify = TRUE,
  header = "> "
)

runFunction.JSON(cmd)
}
\arguments{
\item{what}{either a function or a non-empty character string naming the
    function to be called.}

\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{package}{The name of the package holding the function given by \code{what}.}

\item{removeCall}{Logical: If FALSE, keep the call in the error message.}

\item{onlyStoxMessages}{Logical: If TRUE show only the StoX messages, which are those starting with "StoX: ".}

\item{onlyStoxWarnings}{Logical: If TRUE show only the StoX warnings, which are those starting with "StoX: ".}

\item{onlyStoxErrors}{Logical: If TRUE show only the StoX errors, which are those starting with "StoX: ".}

\item{maxLength.Message, maxLength.Warning, maxLength.Error}{The maximum number of characters to return for messages, warnings and errors, respectively (with allowed values 100...8170, default 1000).}

\item{nwarnings}{The number of warnings to save.}

\item{collapseMessages, collapseWarnings, collapseErrors}{Logical: If TRUE report only unique messages, warnings and errors, respectively, adding the number of occurrences in parenthesis.}

\item{messageLogFile, warningLogFile, errorLogFile}{If given, messages, warnings and errors are printed to these files.}

\item{messageAppend, warningAppend, errorAppend}{Logical: If TRUE, append to the log files (\code{messageLogFile}, \code{warningLogFile} or \code{errorLogFile}, respectively)}

\item{uniquify}{Logical: If TRUE (the default) return unique messages, warnings and errors, with the nnumber of occurrences in parenthesis.}

\item{header}{The string to start each line with.}

\item{cmd}{A JSON string containing parameters listed above.}
}
\description{
\code{runFunction} runs a function using the \code{\link[base]{do.call}} syntax, whereas \code{runFunction.JSON} accepts a JSON string conntaining the parameters to pass on to \code{runFunction}.
}
