% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{runProcess}
\alias{runProcess}
\title{Run a single process:}
\usage{
runProcess(
  projectPath,
  modelName,
  processID,
  msg = TRUE,
  save = TRUE,
  replaceArgs = list()
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{msg}{Logical: If TRUE print a message that tells that the process is being run.}

\item{save}{Logical: If TRUE save the project after the process has been run.}

\item{replaceArgs}{A list of arguments to override the arguments of the process by.}
}
\description{
Run a single process:
}
