% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{runProcesses}
\alias{runProcesses}
\title{Run processes of a StoX model.}
\usage{
runProcesses(
  projectPath,
  modelName,
  startProcess = 1,
  endProcess = Inf,
  save = TRUE,
  force.restart = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{startProcess}{The process index of the first process to run.}

\item{endProcess}{The process index of the last process to run.}

\item{save}{Logical: If TRUE save the project after the process has been run.}

\item{force.restart}{Logical: If TRUE restart the model even if it is running (e.g. if the model crashed before completing).}
}
\description{
This function runs the processes with process index form startProcess to endProcess.
}
