% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{runProject}
\alias{runProject}
\title{Run all models of a StoX project}
\usage{
runProject(
  projectPath,
  run = TRUE,
  save = TRUE,
  force.restart = FALSE,
  replaceDataList = list(),
  replaceArgs = list(),
  fileOutput = NULL,
  setUseProcessDataToTRUE = TRUE,
  purge.processData = FALSE,
  try = TRUE,
  ...
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{run}{Logical: If TRUE run the model.}

\item{save}{Logical: If TRUE save the project after running.}

\item{force.restart}{Logical: If TRUE restart the model before running.}
}
\value{
A list of model output.
}
\description{
This function runs and returns output from all models of a StoX project.
}
