% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{runProject_ReplaceAcousticFiles}
\alias{runProject_ReplaceAcousticFiles}
\title{Add a process DefineAcousticPSU with DefinitionMethod "PreDefined"}
\usage{
runProject_ReplaceAcousticFiles(
  projectPath,
  ReadAcoustic.FileNames,
  modelNames = getRstoxFrameworkDefinitions("stoxModelNames"),
  projectPath0 = NULL,
  run = TRUE,
  AcousticPSUProcessName = "DefineAcousticPSU",
  AcousticPSUProcessNameNew = "DefineAcousticPSU0",
  save = NULL,
  ow = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{ReadAcoustic.FileNames}{Optional: character vector with file paths to input acoustic files to replaec the existing acoustic files by in the procecss using the function \code{ReadAcoustic}.}

\item{modelNames}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}

\item{projectPath0}{Optional: Path to the project to copy to the project given by \code{projectPath}.}

\item{run}{Logical: If TRUE run the model.}

\item{AcousticPSUProcessName}{The name of the existing process using DefineAcousticPSU.}

\item{AcousticPSUProcessNameNew}{The new name of the existing process using DefineAcousticPSU.}

\item{save}{Logical: If TRUE save the project after running.}

\item{ow}{Logical: If TRUE overwrite the project.}
}
\value{
A list of model output.
}
\description{
This function adds a process named DefineAcousticPSU using the function DefineAcousticPSU with DefinitionMethod = "PreDefined". It is a requirement that there exists a procecss named DefineAcousticPSU in the project, which will be renamed and used as input to the new DefineAcousticPSU process.
}
