% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{runProjects}
\alias{runProjects}
\title{Run all models of a StoX project}
\usage{
runProjects(
  projectPaths,
  modelNames = getRstoxFrameworkDefinitions("stoxModelNames"),
  processes = NULL,
  startProcess = 1,
  endProcess = Inf,
  run = TRUE,
  save = TRUE,
  force.restart = FALSE,
  replaceDataList = list(),
  replaceArgs = list(),
  fileOutput = NULL,
  setUseProcessDataToTRUE = TRUE,
  purge.processData = FALSE,
  try = TRUE,
  drop = TRUE,
  close = FALSE,
  ...
)
}
\arguments{
\item{startProcess}{xx}

\item{endProcess}{Integer value speicfying the}

\item{run}{Logical: If TRUE run the model.}

\item{save}{Logical: If TRUE save the project after running.}

\item{force.restart}{Logical: If TRUE restart the model before running.}
}
\value{
A list of model output.
}
\description{
This function runs and returns output from all models of a StoX project.
}
