% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectIO.R
\name{saveProjectEdvin}
\alias{saveProjectEdvin}
\title{Save Stox project to project xml file using namespace http://www.imr.no/formats/stox/v3}
\usage{
saveProjectEdvin(projectDescription, filename)
}
\arguments{
\item{projectDescription}{nested list representation of StoX project}

\item{filename}{to write to}
}
\description{
Save Stox project to project xml file using namespace http://www.imr.no/formats/stox/v3
}
