% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes.R
\name{setProcessPropertyValue}
\alias{setProcessPropertyValue}
\title{GUI function: Set procecss properties.}
\usage{
setProcessPropertyValue(
  groupName,
  name,
  value,
  projectPath,
  modelName,
  processID
)
}
\arguments{
\item{groupName}{The name of the property group, one of "processArguments", "functionInputs" and "functionParameters".}

\item{name}{The name of the property, such as "processName", "functionName", one of the process parameters ("enabled", "showInMap" and "fileOutput"), the name of a funciton input, or the name of a function parameter.}

\item{value}{The value to set to the property (string).}

\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}
}
\description{
GUI function: Set procecss properties.
}
