% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{setUseProcessDataToFALSE}
\alias{setUseProcessDataToFALSE}
\title{Convenience function to set UseProcessData to FALSE for a process.}
\usage{
setUseProcessDataToFALSE(projectPath, modelName, processID)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}
}
\description{
This function cna be used if one wishes to run a process with UseProcessData turned off, thus overriding the process data. UseProcessData is set to TRUE automatically when a process is run, so \code{setUseProcessDataToFALSE} corresponds to unchecking that parameter in the GUI.
}
