% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Biotic.R
\name{StationLengthDist}
\alias{StationLengthDist}
\title{Station length distribution}
\usage{
StationLengthDist()
}
\arguments{
\item{parameterName}{Parameter descrption.}
}
\value{
A data.table is returned with awesome stuff.
}
\description{
This function calculates length frequency distribution per species per station, either given as percentages, or as counts, possibly normalized by towing distance.
}
\details{
This function is awesome and does excellent stuff.
}
\examples{
x <- 1

}
\seealso{
\code{\link[roxygen2]{roxygenize}} is used to generate the documentation.
}
