% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{stopBootstrap}
\alias{stopBootstrap}
\title{Stop a bootstrap run.}
\usage{
stopBootstrap(projectPath)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}
}
\description{
Stop a bootstrap run.
}
