% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{stopProcesses}
\alias{stopProcesses}
\title{Stop runProcesses().}
\usage{
stopProcesses(projectPath, modelName)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}
}
\description{
Stop runProcesses().
}
