% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectIO.R
\name{stox27project}
\alias{stox27project}
\title{Stox 2.7 project}
\description{
Nested list representation of Stox 2.7 projects.
}
\details{
Naming and definitions follow the XML.
Stox 2.7 projects conforms roughly to the namespace http://www.imr.no/formats/stox/v1.2,
defined at http://www.imr.no/formats/stox/v1.2/stoxv1_2.xsd and in formats/stoxv2.7.xsd.

This schema is used with some additions in the representation, with the data type mapping:
\describe{
 \item{xs:string}{character}
 \item{xs:boolean}{logical}
}

In addition to structures defined in this schema, stox 2.7 may store the following structures:
\describe{
 \item{platformfactor}{list of lists with members 'value', 'covariatesourcetype', 'covariate'.}
 \item{covparam}{list of lists with members 'value', 'covariatetable', 'parameter'}
}
}
