% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{Stratum}
\alias{Stratum}
\alias{addStratum}
\alias{removeStratum}
\alias{modifyStratum}
\title{Stratum manipulation}
\usage{
addStratum(stratum, projectPath, modelName, processID)

removeStratum(stratumName, projectPath, modelName, processID)

modifyStratum(stratum, projectPath, modelName, processID)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}
}
\description{
The functions \code{addStations} and \code{removeStations} adds or removes biotic stations from the Assignment process data of the specified process.
}
\details{
The assignment IDs are refreshed for every change, after sorting the assignemnts by the PSU column.
}
