% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RstoxTesting.R
\name{subsetByAllEqual}
\alias{subsetByAllEqual}
\title{Remove columns where all values are equal both in the x and y element of the input}
\usage{
subsetByAllEqual(xy, keys)
}
\arguments{
\item{xy}{A list of tables \code{x} and \code{y}.}

\item{keys}{The keys of the tables in \code{xy}.}
}
\description{
Remove columns where all values are equal both in the x and y element of the input
}
