% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{unlistToDataType}
\alias{unlistToDataType}
\title{Function to flatten a list of output datat and add names from the levels of the list}
\usage{
unlistToDataType(processOutput)
}
\arguments{
\item{processOutput}{A list of StoX output data.}
}
\description{
Function to flatten a list of output datat and add names from the levels of the list
}
