% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RstoxTesting.R
\name{unzipAndCopyProject}
\alias{unzipAndCopyProject}
\title{Copy a StoX project, after unzipping if zipped}
\usage{
unzipAndCopyProject(projectPath, newProjectPath, copydir = tempdir(), ...)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{newProjectPath}{The path to the copied StoX project.}

\item{copydir}{The directory in which to place the copy, defaulted to the tempdir.}

\item{...}{Passed to \code{\link{copyProject}}.}
}
\value{
The path to the copy of the StoX project.
}
\description{
This function is useful to make a copy of a potentially zipped project.
}
