% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{unzipProject}
\alias{unzipProject}
\title{Function for unzipping a zipped StoX project}
\usage{
unzipProject(projectPath, exdir = ".")
}
\arguments{
\item{projectPath}{The project to be run and tested against the existing output files of the project gievn by \code{projectPath_original}.}

\item{exdir}{The direcory to unzip to, defaulted to the current directory.}
}
\description{
Function for unzipping a zipped StoX project
}
