% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{write_list_as_tables_NetCDFF4}
\alias{write_list_as_tables_NetCDFF4}
\title{Write/append a table/list of tables to NetCDF4}
\usage{
write_list_as_tables_NetCDFF4(
  list,
  filePath,
  nc,
  index,
  dims,
  nchars,
  append = FALSE,
  ow = FALSE,
  missval = -9,
  compression = NA,
  verbose = FALSE,
  validOutputDataClasses = getRstoxFrameworkDefinitions("validOutputDataClasses")
)
}
\arguments{
\item{list}{A list of tables to write to NetCDF4.}

\item{filePath}{The path to the file.}

\item{nc}{A netCDF4 object, overiding the \code{filePath}.}

\item{index}{The index of the current data to write, with number of rows indicated in the \code{dims} argument.}

\item{dims}{A list of lists of dimensions of the tables to write.}

\item{nchars}{A list of lists of lists holding the maximum number of characters for each variable of each table of each process.}

\item{append}{Logical: If TRUE, append to the existing file.}

\item{ow}{Logical: If TRUE, overwrite the existing file.}

\item{missval}{The value to use for representing missing values (NA) for numeric variables.}

\item{compression}{If set to an integer between 1 (least compression) and 9 (most compression), this
 enables compression for the variable as it is written to the file.
 Turning compression on forces the created file to be in netcdf version 4 format,
 which will not be compatible with older software that only reads netcdf version 3 files.}

\item{verbose}{Print debugging information.}

\item{validOutputDataClasses}{A vector of valid output data classes, indicating when to stop the unlisting.}
}
\description{
Write/append a table/list of tables to NetCDF4
}
