% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{writeActiveProcessID}
\alias{writeActiveProcessID}
\title{This function has the primary use of writing the proided activeProcessID to the activeProcessID.txt file}
\usage{
writeActiveProcessID(
  projectPath,
  modelName,
  activeProcessID = NULL,
  affectLaterModels = FALSE,
  currentActiveProcessID = NULL,
  processDirty = NULL,
  propertyDirty = NULL
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{activeProcessID}{The process ID of the process to set as the active process. If not given, the active process ID is not changed.}

\item{affectLaterModels}{Logical: If TRUE reset the later models it the active pocess ID has changed in the current model.}

\item{currentActiveProcessID}{If both activeProcessID and currentActiveProcessID, a check is made for whether these are equal, in which case the active process ID is NOT written.}

\item{processDirty}{Logical: Controls whether the process has been changed (either process data or parameters) by user and not run.}

\item{propertyDirty}{Logical: Controls whether the properties of the process has been changed by a run, i.e. UseProcessdata automatically set to TRUE}
}
\description{
This function has the primary use of writing the proided activeProcessID to the activeProcessID.txt file
}
