% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{writeProjectDescription}
\alias{writeProjectDescription}
\title{Write the project description.}
\usage{
writeProjectDescription(
  projectPath,
  projectDescription = NULL,
  projectDescriptionFile = NULL,
  Application = R.version.string,
  optionalDependencies = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{projectDescription}{The project description to write to the project.json file. Defaulted to NULL, in which case the project description is read for the project.}

\item{projectDescriptionFile}{The path to the file holding the projectDescription.}

\item{Application}{The application running StoX. Defaulted to the R.version.string, but can be a StoX version when run from the GUI.}

\item{optionalDependencies}{Logical: If TRUE include also the column OptionalDependencies from the OfficialRstoxFrameworkVersions.txt file, which holds Rstox packages used in the Suggests field of the DESCRIPTION file.}

\item{verbose}{Logical: If TRUE extra messages are printed to console.}
}
\description{
Write the project description.
}
