% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{writeProjectDescriptionJSON}
\alias{writeProjectDescriptionJSON}
\title{Write project description to a file in json format}
\usage{
writeProjectDescriptionJSON(projectDescription, projectDescriptionFile, GUI)
}
\arguments{
\item{projectDescription}{a list of lists with project description.}

\item{projectDescriptionFile}{a file name.}
}
\description{
This function writes project description to json file.
}
