% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{writeStoxJsonSchema}
\alias{writeStoxJsonSchema}
\title{Export StoX JSON schema}
\usage{
writeStoxJsonSchema(con)
}
\arguments{
\item{con}{A connection to which to write the schema. Returned as JSON if missing.}
}
\description{
Export StoX JSON schema
}
