% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{zipProject}
\alias{zipProject}
\title{Zip a StoX project}
\usage{
zipProject(projectPath, zipPath = NULL, zipDir = NULL, ow = FALSE)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{zipPath}{The file path of the ziped StoX project.}

\item{zipDir}{The file path of the directory in which to place the ziped StoX project.}

\item{ow}{Logical: If TRUE overwrite the zip if already present.}
}
\description{
Zip a StoX project
}
