% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{AcousticPSU}
\alias{AcousticPSU}
\alias{addAcousticPSU}
\alias{removeAcousticPSU}
\alias{addEDSU}
\alias{removeEDSU}
\title{Add, remove and renamme acoustic PSUs and EDSUs}
\usage{
addAcousticPSU(Stratum, PSU = NULL, projectPath, modelName, processID)

removeAcousticPSU(PSU, projectPath, modelName, processID)

addEDSU(PSU, EDSU, projectPath, modelName, processID)

removeEDSU(EDSU, projectPath, modelName, processID)
}
\arguments{
\item{Stratum}{The name of a stratum.}

\item{PSU}{The name of a PSU.}

\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{EDSU}{The name of an EDSU.}
}
\description{
The functions \code{addAcousticPSU}, \code{removeAcousticPSU} and \code{renameAcousticPSU} adds, removes or renames AcousticPSU from the AcousticPSU process data of the specified process. 
The functions \code{addEDSU} and \code{removeEDSU} adds or removes EDSUSs from the AcousticPSU process data of the specified process. 
The function \code{removeAllAcousticPSUsOfStratum} removes all PSUs from the AcousticPSU process data of the specified process.
}
\details{
The assignment IDs are refreshed for every change, after sorting the assignemnts by the PSU column.
}
