% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interactive.R
\name{Assignment}
\alias{Assignment}
\alias{addHaulToAssignment}
\alias{removeHaulFromAssignment}
\title{Add or remove biotic hauls from assignments}
\usage{
addHaulToAssignment(PSU, Layer, Haul, projectPath, modelName, processID)

removeHaulFromAssignment(PSU, Layer, Haul, projectPath, modelName,
  processID)
}
\arguments{
\item{PSU}{The acoustic primary sampling unit (PSU) for which to remove the haul}

\item{Layer}{The acoustic Layer for which to remove the haul}

\item{Haul}{The biotic haul to remove (can be a vector of hauls).}

\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}
}
\description{
The functions \code{addHaulsToAssignment} and \code{removeHaulsFromAssignment} adds or removes biotic hauls from the Assignment process data of the specified process.
}
\details{
The assignment IDs are refreshed for every change, after sorting the assignemnts by the PSU column.
}
\examples{
## # Create artificial assignment data:
## Assignment <- data.table::data.table(
##     PSU = paste0("T", c(1,1, 2, 3,3)), 
##     Layer = "L1", 
##     Haul = c(1,2, 2, 1,2), 
##     AssignmentID = c(1,1, 2, 1,1)
## )
## 
## # Add a haul:
## Assignment1 <- assignment_addHauls("T3", "L1", 4, Assignment)
## all.equal(Assignment, Assignment2)
## 
## # Remove the same haul:
## Assignment2 <- assignment_removeHauls("T3", "L1", 4, Assignment1)
## all.equal(Assignment, Assignment2)

}
