% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Bootstrap the baseline model}
\usage{
Bootstrap(
  outputData,
  projectPath,
  BootstrapMethodTable = data.table::data.table(),
  NumberOfBootstraps = 1L,
  OutputProcesses = character(),
  UseOutputData = FALSE,
  NumberOfCores = 1L,
  BaselineSeedTable = data.table::data.table()
)
}
\arguments{
\item{outputData}{The output of the function from an earlier run.}

\item{projectPath}{The path to the project to containing the baseline to bootstrap.}

\item{BootstrapMethodTable}{A table of the columns ProcessName, ResampleFunction and Seed, where each row defines the resample function to apply to the output of the given process (and the seed to use in the resampling). Run RstoxFramework::getResampleFunctions() to get a list of the implemented resample functions. Note that if a process is selected inn \code{BootstrapMethodTable} that is not used in the model up to the \code{OutputProcesses}, the bootstrapping of that process will not be effective on the end result (e.g. select the correct process that returns BioticAssignment data type).}

\item{NumberOfBootstraps}{Integer: The number of bootstrap replicates.}

\item{OutputProcesses}{A vector of the processes to save from each bootstrap replicate.}

\item{UseOutputData}{Logical: Bootstrapping can be time consuming, and by setting \code{UseOutputData} to TRUE the output file generated by a previous run of the process will be used instead of re-running the bootstrapping. Use this parameter with caution, as it any changes made to the Baseline model after the previous bootstrap run will not be included unless one selects \code{UseOutputData} = FALSE.}

\item{NumberOfCores}{The number of cores to use for parallel processing. A copy of the project is created in tempdir() for each core, also when using only one core. Note that this will require disc space equivalent to the \code{NumberOfCores} time the size of the project folder (excluding the output/analysis/Bootstrap folder, which will be deleted before copies are made).}

\item{BaselineSeedTable}{A table of ProcessName and Seed, givng the seed to use for the Baseline processes that requires a Seed parameter.}
}
\value{
A \code{\link{BootstrapData}} object, which is a list of the RstoxData \code{\link[RstoxData]{DataTypes}} and RstoxBase \code{\link[RstoxBase]{DataTypes}}.
}
\description{
Run a subset of the baseline model a number of times after resampling e.g. Hauls in each Stratum, EDUSs in each Stratum.
}
\details{
A copy of the project is made for each core given by \code{NumberOfCores}. In the case that NumberOfCores == 1, this is still done for safety.
}
