% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Bootstrap the baseline model}
\usage{
Bootstrap(
  projectPath,
  BootstrapMethodTable = data.table::data.table(),
  NumberOfBootstraps = 1L,
  OutputProcesses = character()
)
}
\arguments{
\item{LengthDistributionData}{The \code{\link[RstoxBase]{LengthDistributionData}} to resample Hauls in.}
}
\description{
Run a subset of the baseline model a number of times after resampling e.g. Hauls in each Stratum, EDUSs in each Stratum.
}
