% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{BootstrapData}
\alias{BootstrapData}
\title{Bootstrap data}
\description{
The BootstrapData model data is a list of outputs from the baseline generated by re-running baseline with random sampling of the PSUs and assignments (in the case of acoustic-trawl estimates).
}
\seealso{
This data type is produced by the function \code{\link{Bootstrap}}. See \code{\link{DataTypes}} for a list of all data types produced by \code{\link{RstoxData}}, \code{\link{RstoxBase}} and \code{\link{RstoxFramework}}.
}
