% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{compareProjectToStoredOutputFiles}
\alias{compareProjectToStoredOutputFiles}
\title{Function for comparing existing output files with the memory read using runProject()}
\usage{
compareProjectToStoredOutputFiles(
  projectPath,
  projectPath_original = projectPath
)
}
\arguments{
\item{projectPath}{The project to be run and tested against the existing output files of the project gievn by \code{projectPath_original}.}

\item{projectPath_original}{The project holding the existing output files, defaulted to \code{projectPath}.}
}
\description{
Function for comparing existing output files with the memory read using runProject()
}
