% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{convertStoX2.7To3}
\alias{convertStoX2.7To3}
\title{Backward compabitibility actions:}
\usage{
convertStoX2.7To3(
  projectPath2.7,
  projectPath3,
  newProjectPath3 = NULL,
  ow = FALSE
)
}
\arguments{
\item{projectPath2.7}{The path to the StoX 2.7 project to convert to StoX 3.x.x.}

\item{projectPath3}{The path to the StoX 3.x.x project used as model for the conversion.}

\item{newProjectPath3}{The path to the StoX 3.x.x project to create.}

\item{ow}{Logical: If TRUE overwrite the project.}
}
\description{
Backward compabitibility actions:
}
