% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{convertStoX2.7To3}
\alias{convertStoX2.7To3}
\title{Backward compabitibility actions:}
\usage{
convertStoX2.7To3(
  projectPath2.7,
  projectPath3,
  newProjectPath3,
  functionsToApplyProjectXML2.7To = c("DefineSurvey", "DefineStratumPolygon",
    "DefineBioticPSU", "DefineAcousticPSU", "DefineBioticAssignment"),
  ow = FALSE,
  run = TRUE,
  msg = FALSE
)
}
\arguments{
\item{projectPath2.7}{The path to the StoX 2.7 project to convert to StoX 3.x.x.}

\item{projectPath3}{The path to the StoX 3.x.x project used as model for the conversion.}

\item{newProjectPath3}{The path to the StoX 3.x.x project to create.}

\item{functionsToApplyProjectXML2.7To}{A character vector of names of process data functions to read from the project.xml of the \code{projectPath2.7}.}

\item{ow}{Logical: If TRUE, overwrite the \code{newProjectPath3}.}

\item{run}{Logical: If TRUE, run the entire project after converting. If FALSE, changes will not be stored in the project.json before one opens and runs the baseline.}

\item{msg}{Logical: If TRUE, print extra messages.}
}
\description{
Backward compabitibility actions:
}
