% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{formatProcess}
\alias{formatProcess}
\title{Function to format a process as read from the project.json}
\usage{
formatProcess(process)
}
\arguments{
\item{parameter}{}

\item{simplifyVector}{}
}
\description{
Function to format a process as read from the project.json
}
