% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{general_arguments}
\alias{general_arguments}
\title{General parameters of RstoxFramework.}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{projectPaths}{A vector of paths to the StoX projects.}

\item{projectDescriptionFile}{The path to the file holding the projectDescription (project.json).}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{modelNames}{The name of the models (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{processName}{The name of the process.}

\item{processNames}{A vector of names of the processes.}

\item{processes}{The name of the processes to return data from.}

\item{functionName}{The name of the function used by the process.}

\item{tableName}{The name of the output table to get from the process.}

\item{startProcess}{The process index, name or ID at which to start the model run.}

\item{endProcess}{The process index, name or ID at which to stop the model run.}

\item{beforeProcessID}{The ID of the process before which to get the process table.}

\item{afterProcessID}{The ID of the process after which to get the process table or to place a process.}

\item{warn}{Logical: If TRUE show warnings that are not highly important.}

\item{verbose}{Logical: If TRUE extra messages are printed to console.}

\item{msg}{Logical: If FALSE no messages are printed to console (except possibly for extremely important ones).}

\item{msg.GUI}{Logical: If TRUE, exlcude certain messages (used by the StoX GUI).}

\item{archive}{Logical: Should the project memory state be archived using \code{archiveProject}? This stores one step in the history.}

\item{template}{A string naming the template to use when generating the project. See \code{getAvaiableTemplates} for a list of available templates.}

\item{ow}{Logical: If TRUE overwrite the project.}

\item{showWarnings}{Logical: If TRUE display warninigs when creting the project folders.}

\item{Application}{A single string naming the application used when saving the project. Defaulted to R.version.string.}

\item{argumentFilePaths}{A nested list of paths to argument files of a model, as returned from \code{getArgumentFilePaths}. This is used to speed up some functions.}

\item{only.valid}{Logical: If TRUE subset function arguments (inputs and parameters) to only those to be shown as a consequence of argument hierarchy (e.g., one argument being irrelevant for a specific setting of another).}

\item{returnProcessTable}{Logical: If TRUE return the process table (much used in functions used by the GUI).}

\item{applyBackwardCompatibility}{Logical: If TRUE apply backward compatibility actions when running \code{readProjectDescription}.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
All functions referring to a project, a model, a process or an output table use the same parameters, listed here.
}
