% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getAvailableStoxFunctionNames}
\alias{getAvailableStoxFunctionNames}
\title{Function returning the names of the StoX functions available for a model:}
\usage{
getAvailableStoxFunctionNames(modelName)
}
\arguments{
\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}
}
\description{
Function returning the names of the StoX functions available for a model:
}
