% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{getBootstrapData}
\alias{getBootstrapData}
\title{Get bootstrap data saved in a NetCDF4 file}
\usage{
getBootstrapData(
  filePath,
  nc,
  selection = list(),
  BootstrapIDStart = 1,
  BootstrapIDEnd = Inf,
  unlist = 0,
  close = TRUE
)
}
\arguments{
\item{filePath}{The path to the file.}

\item{nc}{A netCDF4 object, overiding the \code{filePath}.}

\item{selection}{Hierarchical list of names of the groups/variables. The last element must be a vector of the variables to return from the table specified by the other elements. E.g., list("ImputeSuperIndividuals", "SuperIndividualsData", c("Stratum", "IndividualAge", "Abundance")) will return a data.table of the three columns "Stratum", "IndividualAge" and "Abundance", added the BootstrapID specified in \code{BootstrapID}.}

\item{BootstrapIDStart, BootstrapIDEnd}{The start and end bootstrap IDs, i.e., the indices of the bootstrap replicates. The default returns all bootstrap replicates.}

\item{unlist}{For use when a single table is requested. If \code{unlist} = 0 (the default) return a list named by the \code{selection} with a sub-list named by the datatype holding the output data  (e.g. $ImputeSuperIndividuals$SuperIndividualsData). If \code{unlist} = 1 return a list named by the \code{selection} holding the output data. If \code{unlist} = 2 return the output data directly. Use \code{unlist} = 1 to replicate the output of Bootstrap() in StoX <= 3.6.2.}

\item{close}{Logical: If TRUE (the default) close the netCDF4 file after reading the data.}
}
\description{
Get bootstrap data saved in a NetCDF4 file
}
