% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Routes.R
\name{getFilterOptions}
\alias{getFilterOptions}
\title{GUI function: Get possible tables, operators and unique values for use in the filter expression builder.}
\usage{
getFilterOptions(projectPath, modelName, processID, tableName)
}
\description{
GUI function: Get possible tables, operators and unique values for use in the filter expression builder.
}
