% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getModelData}
\alias{getModelData}
\title{Get output data from processes of a StoX model}
\usage{
getModelData(
  projectPath,
  modelName,
  processes = NULL,
  startProcess = 1,
  endProcess = Inf,
  drop.datatype = TRUE,
  warn = TRUE,
  unlistDepth2 = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}

\item{processes}{The name of the process(es).}

\item{startProcess}{The process index, name or ID at which to start and stop the model run.}

\item{endProcess}{The process index, name or ID at which to start and stop the model run.}

\item{drop.datatype}{Logical: If TRUE, drop the top level of the output list if it has length 1 and that single element is named by the datatype name.}

\item{warn}{Logical: If TRUE show warnings that are not highly important.}

\item{unlistDepth2}{Logical: Related to \code{drop.datatype}, but setting this to TRUE unlists output data that are nested in 2 levels, such as output from \code{\link[RstoxData]{ReadBiotic}}, which outputs a set of tables for each input file. Using unlistDepth2 = TRUE puts all these tables in one list, and uses the concatenation of the file names and the table name separated by underscore. This is how it is displayed in the StoX GUI when selecting "View output".}
}
\description{
Get output data from processes of a StoX model
}
