% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{getOfficialRstoxPackageVersion}
\alias{getOfficialRstoxPackageVersion}
\title{Get certified Rstox package versions}
\usage{
getOfficialRstoxPackageVersion(
  StoXVersion = NULL,
  officialRstoxPackageVersionsFile = system.file("versions",
    "OfficialRstoxFrameworkVersions.txt", package = "RstoxFramework"),
  optionalDependencies = FALSE,
  toJSON = FALSE,
  list.out = FALSE
)
}
\arguments{
\item{StoXVersion}{The version of the StoX GUI defining the combination of official Rstox package versions.}

\item{officialRstoxPackageVersionsFile}{The path to the file holding the link between StoX GUI version and Rstox package versions. If missing, the file on the RstoxFramework master on GitHub is used ("https://raw.githubusercontent.com/StoXProject/RstoxFramework/master/inst/versions/OfficialRstoxFrameworkVersions.txt").}

\item{optionalDependencies}{Logical: If TRUE include also the column OptionalDependencies from the OfficialRstoxFrameworkVersions.txt file, which holds Rstox packages used in the Sugggests field of the DESCRIPTION file.}

\item{toJSON}{Logical: If TRUE output a JSON string.}

\item{list.out}{Logical: If TRUE wrap the output of \code{getOfficialRstoxPackageVersion} in a list with packageName and version.}
}
\description{
This function reads the OfficialRstoxFrameworkVersions.txt file for the given StoX GUI version.
}
