% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getProcessTable}
\alias{getProcessTable}
\alias{scanForModelError}
\alias{getProcessesSansProcessData}
\alias{getProcessAndFunctionNames}
\title{Functions to get the process table of a model.}
\usage{
getProcessTable(
  projectPath,
  modelName = NULL,
  afterProcessID = NULL,
  beforeProcessID = NULL,
  argumentFilePaths = NULL,
  only.valid = TRUE
)

scanForModelError(
  projectPath,
  modelName = NULL,
  afterProcessID = NULL,
  beforeProcessID = NULL,
  argumentFilePaths = NULL,
  only.valid = TRUE
)

getProcessesSansProcessData(
  projectPath,
  modelName = NULL,
  afterProcessID = NULL,
  beforeProcessID = NULL,
  argumentFilePaths = NULL,
  only.valid = FALSE
)

getProcessAndFunctionNames(
  projectPath,
  modelName = NULL,
  afterProcessID = NULL,
  beforeProcessID = NULL,
  argumentFilePaths = NULL
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}
}
\description{
Functions to get the process table of a model.
}
