% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getProjectPaths}
\alias{getProjectPaths}
\title{Get paths to the project directories and files}
\usage{
getProjectPaths(ProjectName, ProjectDirectory = NULL)
}
\arguments{
\item{ProjectName}{The directory in which to put the "stox" folder, defaulted to the "workspace" folder in the home directory.}

\item{ProjectDirectory}{The directory in which to put the "stox" folder, defaulted to the "workspace" folder in the home directory.}
}
\value{
A list of paths to the "stox" folder and sub folders.
}
\description{
This function gets the paths to the top level, input and output folders, and the project recipe file.
}
\examples{
getProjectPaths()

}
\seealso{
Use \code{\link{createStoxSkeleton}} to create the folders.
}
