% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{getProjectPaths}
\alias{getProjectPaths}
\title{This function gets the paths defined by \code{\link{initiateRstoxFramework}}.}
\usage{
getProjectPaths(projectPath, name = NULL)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{name}{A string naming the path element to get. Set this to NULL to get all paths.}
}
\description{
This function gets the paths defined by \code{\link{initiateRstoxFramework}}.
}
