% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Versions.R
\name{installOfficialRstoxPackagesWithDependencies}
\alias{installOfficialRstoxPackagesWithDependencies}
\title{Function to download all official Rstox packages and their CRAN dependencies}
\usage{
installOfficialRstoxPackagesWithDependencies(
  StoXGUIVersion,
  officialRstoxPackageVersionsFile,
  destdir = NA,
  Rstox.repos = "https://stoxproject.github.io/repo",
  download.repos = "https://cloud.r-project.org",
  dependencies = NA,
  sort = FALSE,
  platform = NA,
  twoDigitRVersion = NA,
  ...
)
}
\arguments{
\item{StoXGUIVersion}{The version of the StoX GUI defining the combination of official Rstox package versions.}

\item{officialRstoxPackageVersionsFile}{The path to the file holding the link between StoX GUI version and Rstox package versions.}

\item{destdir}{The directory to download binaries to, defaulted to NA which implies tempdir().}

\item{Rstox.repos}{Either NULL to search for packages locally, or a repository passed to \code{\link[utils]{available.packages}}, where \code{Rstox.repos} is used to locate direct dependencies of the Rstox packages, and \code{nonRstox.reposs} is used to get the dependencies of the direct dependencies recursievly, defaulted to the CRAN repository.}

\item{dependencies}{A character vector listing the dependencies to get, with possible values "Depends", "Imports", "LinkingTo", "Suggests", "Enhances". Default is NA, implying c("Depends", "Imports", "LinkingTo").}

\item{sort}{Logical: If TRUE sort the dependent packages, defaulted to FALSE to enable installing the most basic dependencies first.}

\item{platform}{The platfor to downoad binaries for, defaulted to the current platform, but with the option to download for other platforms (possible values are "windows" and "macosx").}

\item{twoDigitRVersion}{The two digit R version to downoad binaries for, defaulted to the current version, but with the option to download for other versions such as 3.6.}
}
\description{
Function to download all official Rstox packages and their CRAN dependencies
}
