% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{modifyProcess}
\alias{modifyProcess}
\title{Modify a process}
\usage{
modifyProcess(
  projectPath,
  modelName,
  processName,
  newValues,
  archive = TRUE,
  add.defaults = FALSE,
  purge.processData = FALSE
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processName}{The name of the process.}
}
\description{
Modify a process
}
