% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{parseParameter}
\alias{parseParameter}
\title{Parse a parameter received from the GUI, usually saved as JSON.}
\usage{
parseParameter(parameter, simplifyVector = TRUE)
}
\arguments{
\item{parameter}{}

\item{simplifyVector}{}
}
\description{
Parse a parameter received from the GUI, usually saved as JSON.
}
