% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{parseParameter}
\alias{parseParameter}
\title{Function to parse a parameter coming from the GUI}
\usage{
parseParameter(parameter, simplifyVector = TRUE)
}
\arguments{
\item{parameter}{A JSON string holding the parameters.}

\item{simplifyVector}{coerce JSON arrays containing only primitives into an atomic vector}
}
\description{
Function to parse a parameter coming from the GUI
}
