% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoxFunctionAttributes.R
\docType{data}
\name{processPropertyFormats}
\alias{processPropertyFormats}
\title{Process property formats for RstoxFramework}
\format{
An object of class \code{list} of length 5.
}
\usage{
processPropertyFormats
}
\description{
Process property formats for RstoxFramework
}
\keyword{datasets}
