% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{Projects}
\alias{Projects}
\alias{createProject}
\alias{openProject}
\alias{closeProject}
\alias{saveProject}
\alias{saveAsProject}
\alias{copyProject}
\title{Create, open, close, save, saveAs and copy StoX projects.}
\usage{
createProject(
  projectPath,
  template = "EmptyTemplate",
  ow = FALSE,
  showWarnings = FALSE,
  open = TRUE
)

openProject(
  projectPath,
  showWarnings = FALSE,
  force = FALSE,
  reset = FALSE,
  type = c("RData", "JSON")
)

closeProject(projectPath, save = NULL, msg = TRUE)

saveProject(projectPath, type = c("RData", "JSON"), force = FALSE)

saveAsProject(projectPath, newProjectPath, ow = FALSE)

copyProject(projectPath, newProjectPath, ow = FALSE)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{template}{A string naming the template to use when generating the project. See \code{getAvaiableTemplates} for a list of available templates.}

\item{ow}{Logical: If TRUE overwrite the project.}

\item{showWarnings}{Logical: If TRUE display warninigs when creting the project folders.}

\item{open}{Logical: If TRUE open the project after creating it.}

\item{force}{Logical: If TRUE reopen (close and then open) the project if already open.}

\item{reset}{Logical: If TRUE reset each model to the start of the model.}

\item{type}{The type of file to save the project to.}

\item{save}{Logical: If TRUE save the project before closing. Default (NULL) is to ask the user whether to save the project before closing.}

\item{newProjectPath}{The path to the copied StoX project.}
}
\description{
Create a StoX project using \code{createProject}, possibly specifying a template; open an existing (un-opened) project using \code{openProject}, which involves creating files holding the memory of the project; close a project using  \code{closeProject}, which removes the memory files; save the project using \code{saveProject}, which saves the memory file to the project description file or make a copy using \code{saveProject} or \code{saveProject}, where the former closes the given project unsaved and opens the copy.
}
