% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{ProjectUtils}
\alias{ProjectUtils}
\alias{isProject}
\alias{isSaved}
\alias{isOpenProject}
\alias{readProjectDescription}
\alias{writeProjectDescription}
\alias{initiateActiveProcessID}
\alias{isRunning}
\alias{setRunning}
\alias{setNotRunning}
\title{Utilities for projects.}
\usage{
isProject(projectPath)

isSaved(projectPath)

isOpenProject(projectPath)

readProjectDescription(
  projectPath,
  type = getRstoxFrameworkDefinitions("projectDescriptionFileFormats"),
  verbose = FALSE,
  projectDescriptionFile = NULL
)

writeProjectDescription(
  projectPath,
  type = c("JSON", "RData"),
  Application = R.version.string,
  projectDescription = NULL,
  verbose = FALSE
)

initiateActiveProcessID(projectPath)

isRunning(projectPath, modelName)

setRunning(projectPath, modelName)

setNotRunning(projectPath, modelName)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{type}{The type of file to read.}

\item{verbose}{Logical: If TRUE, print extra information to the console, e.g. about backward compatibility.}

\item{projectDescriptionFile}{The optional path to the project description (project.json) file.}

\item{Application}{A string naming the application (e.g. graphical user interface such as StoX) used to produce the description file.}

\item{projectDescription}{The projectDescription as read from the \code{projectDescriptionFile}.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}
}
\description{
Utilities for projects.

Read the project description.

Write the project description.

Initiate the actige processID.

Check or set if a project is running or not.
}
