% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{ProjectUtils}
\alias{ProjectUtils}
\alias{isProject}
\alias{isSaved}
\alias{isOpenProject}
\alias{readProjectDescription}
\alias{writeProjectDescription}
\alias{initiateActiveProcessID}
\alias{isRunning}
\alias{setRunning}
\alias{setNotRunning}
\title{Utilities for projects.}
\usage{
isProject(projectPath)

isSaved(projectPath)

isOpenProject(projectPath)

readProjectDescription(projectPath, type = c("RData", "XML"))

writeProjectDescription(projectPath, type = c("RData", "XML"))

initiateActiveProcessID(projectPath)

isRunning(projectPath)

setRunning(projectPath)

setNotRunning(projectPath)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}
}
\description{
Utilities for projects.
}
