% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{ProjectUtils}
\alias{ProjectUtils}
\alias{isProject}
\alias{isSaved}
\alias{isOpenProject}
\alias{readProjectDescription}
\alias{writeProjectDescription}
\alias{initiateActiveProcessID}
\alias{isRunning}
\alias{setRunning}
\alias{setNotRunning}
\title{Utilities for projects.}
\usage{
isProject(projectPath)

isSaved(projectPath)

isOpenProject(projectPath)

readProjectDescription(
  projectPath,
  type = getRstoxFrameworkDefinitions("projectDescriptionFileFormats"),
  verbose = FALSE,
  projectDescriptionFile = NULL,
  applyBackwardCompatibility = TRUE,
  formatProcesses = TRUE
)

writeProjectDescription(
  projectPath,
  type = c("JSON", "RData"),
  Application = R.version.string,
  projectDescription = NULL,
  verbose = FALSE
)

initiateActiveProcessID(projectPath)

isRunning(projectPath, modelName)

setRunning(projectPath, modelName)

setNotRunning(projectPath, modelName)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{type}{The type of file to read.}

\item{verbose}{Logical: If TRUE extra messages are printed to console.}

\item{projectDescriptionFile}{The path to the file holding the projectDescription.}

\item{applyBackwardCompatibility}{Logical: If TRUE apply backward compatibility actions when running \code{readProjectDescription}.}

\item{formatProcesses}{Logical: If TRUE format the processes after reading the projectDescription file, ensuring correct primitive types. This has a use of FALSE in \code{readModelData}, but should otherwise be set to TRUE.}

\item{Application}{A single string naming the application used when saving the project. Defaulted to R.version.string.}

\item{projectDescription}{A list holding the project description, as read using \code{\link{readProjectDescription}}.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}
}
\description{
Utilities for projects.

Read the project description.

Write the project description.

Initiate the actige processID.

Check or set if a project is running or not.
}
