% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{ProjectUtils}
\alias{ProjectUtils}
\alias{isProject}
\alias{isSaved}
\alias{isOpenProject}
\alias{readProjectDescription}
\alias{writeProjectDescription}
\alias{initiateActiveProcessID}
\alias{isRunning}
\alias{setRunning}
\alias{setNotRunning}
\title{Utilities for projects.}
\usage{
isProject(projectPath)

isSaved(projectPath)

isOpenProject(projectPath)

readProjectDescription(
  projectPath,
  type = getRstoxFrameworkDefinitions("projectDescriptionFileFormats")
)

writeProjectDescription(
  projectPath,
  type = c("JSON", "RData"),
  GUI = "unspecified"
)

initiateActiveProcessID(projectPath)

isRunning(projectPath, modelName)

setRunning(projectPath, modelName)

setNotRunning(projectPath, modelName)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{type}{The type of file to read.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}
}
\description{
Utilities for projects.

Read the project description.

Write the project description.

Initiate the actige processID.

Check or set if a project is running or not.
}
