% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{ProjectUtils}
\alias{ProjectUtils}
\alias{isProject}
\alias{isSaved}
\alias{isOpenProject}
\alias{readProjectDescription}
\alias{initiateActiveProcessID}
\alias{isRunning}
\alias{setRunning}
\alias{setNotRunning}
\title{Utilities for projects.}
\usage{
isProject(projectPath)

isSaved(projectPath)

isOpenProject(projectPath, strict = FALSE)

readProjectDescription(
  projectPath,
  verbose = FALSE,
  projectDescriptionFile = NULL,
  applyBackwardCompatibility = TRUE,
  formatProcesses = TRUE,
  validateJSON = TRUE
)

initiateActiveProcessID(projectPath)

isRunning(projectPath, modelName)

setRunning(projectPath, modelName)

setNotRunning(projectPath, modelName)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{strict}{Logical: If TRUE, require that all folders of the projectSession folder exist in isOpenProject(). Otherwise only require that the projectSession folder exists.}

\item{verbose}{Logical: If TRUE extra messages are printed to console.}

\item{projectDescriptionFile}{The path to the file holding the projectDescription (project.json).}

\item{applyBackwardCompatibility}{Logical: If TRUE apply backward compatibility actions when running \code{readProjectDescription}.}

\item{formatProcesses}{Logical: If TRUE format the processes after reading the projectDescription file, ensuring correct primitive types. This has a use of FALSE in \code{readModelData}, but should otherwise be set to TRUE.}

\item{validateJSON}{Logical: If TRUE validate the project.json.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}
}
\description{
Utilities for projects.

Read the project description.

Initiate the actige processID.

Check or set if a project is running or not.
}
