% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwardCompatibility.R
\name{readStoX2.7OutputFile}
\alias{readStoX2.7OutputFile}
\title{Read output file from StoX 2.7}
\usage{
readStoX2.7OutputFile(x)
}
\arguments{
\item{x}{The path to the StoX 2.7 project file to read.}
}
\description{
Read output file from StoX 2.7
}
