% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{rearrangeProcesses}
\alias{rearrangeProcesses}
\title{Rearrange processes of a StoX model.}
\usage{
rearrangeProcesses(projectPath, modelName, processID, afterProcessID = NULL)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model(s) (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{afterProcessID}{The ID of the process after or before which to get the procecss table (place a process or in the case of afterProcessID).}
}
\description{
Rearrange processes of a StoX model.
}
