% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ReportBootstrap}
\alias{ReportBootstrap}
\title{Report Bootstrap}
\usage{
ReportBootstrap(
  BootstrapData,
  BaselineProcess = character(),
  TargetVariable = character(),
  AggregationFunction = RstoxBase::getReportFunctions(getMultiple = FALSE),
  BootstrapReportFunction = RstoxBase::getReportFunctions(getMultiple = TRUE),
  GroupingVariables = character(),
  RemoveMissingValues = FALSE,
  AggregationWeightingVariable = character(),
  BootstrapReportWeightingVariable = character()
)
}
\arguments{
\item{BootstrapData}{The \code{\link{BootstrapData}} data output from \code{\link{Bootstrap}}.}

\item{BaselineProcess}{A strings naming the baseline process to report from the boostrap output.}

\item{TargetVariable}{The variable to report.}

\item{AggregationFunction}{The function to apply to each bootstrap run. This must be a function returning a single value.}

\item{BootstrapReportFunction}{The function to apply across bootstrap run, such as "cv" or "stoxSummary".}

\item{GroupingVariables}{The variables to report by. For most applications \code{GroupingVariables} should include "Survey" or "SpeciesCategory", unless the user needs to sum over all Survey and SpeciesCategory.}

\item{RemoveMissingValues}{Logical: If TRUE, remove missing values (NAs). The default (FALSE) implies to report NA if at least one of the values used in the \code{ReportFunction} is NA. Use \code{RemoveMissingValues} = TRUE with extreme caution, as it may lead to under-estimation. E.g., if \code{RemoveMissingValues} = TRUE and a super-individual lacks \code{IndividualRoundWeight}, \code{Biomass} will be NA, and the portion of \code{Abundance} distributed to that super-individual will be excluded when summing \code{Biomass} (but included when summing \code{Abundance}). It is advised to always run with \code{RemoveMissingValues} = FALSE first, and make a thorough investigation to identify the source of any missing values. The function \code{link{ImputeSuperIndividuals}} can be used to impute the missing information from other super-individuals.}

\item{AggregationWeightingVariable}{The variable to weight by in the \code{AggregationFunction}.}

\item{BootstrapReportWeightingVariable}{The variable to weight by in the \code{BootstrapReportFunction}.}
}
\value{
A \code{\link{ReportBootstrapData}} object.
}
\description{
Reports the sum, mean or other functions on a variable of the \code{\link{BootstrapData}}.
}
\details{
This function is useful to, e.g, sum Biomass for each SpeciesCategory and IndividualTotalLenght, or average IndividualTotalLenght for each IndiivdualAge and Stratum.
}
