% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleBioticPSUs}
\alias{ResampleBioticPSUs}
\title{Resamples biotic PSUs}
\usage{
ResampleBioticPSUs(MeanLengthDistributionData, Seed, ResampleBy = "Stratum")
}
\arguments{
\item{LengthDistributionData}{The \code{\link[RstoxBase]{LengthDistributionData}} to resample Hauls in.}
}
\description{
This function resamples PSUs with replacement by altering the LengthDistributionWeight.
}
