% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleMeanNASCData}
\alias{ResampleMeanNASCData}
\title{Resamples acoustic PSUs}
\usage{
ResampleMeanNASCData(MeanNASCData, Seed)
}
\arguments{
\item{Seed}{The seed, given as a sinigle initeger.}

\item{MeanNASC}{The \code{\link[RstoxBase]{MeanNASC}} data.}
}
\description{
This function resamples acoustic PSUs with replacement within each Stratum, changing the MeanNASC
}
