% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleMeanNASCData}
\alias{ResampleMeanNASCData}
\title{Resamples acoustic PSUs}
\usage{
ResampleMeanNASCData(MeanNASCData, Seed)
}
\arguments{
\item{MeanNASCData}{The \code{\link[RstoxBase]{MeanNASCData}} data.}

\item{Seed}{The seed, given as a single integer.}
}
\description{
This function resamples, with replacement, the AcousticPSUs within Stratum in the \code{\link{MeanNASCData}}, where Stratum is the stratum associated to the PSU, and not necessarily the actual stratum polygon. The column NASC is scaled by the number of occurrences of each AcousticPSUs from the resampling.
}
