% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{ResampleMeanSpeciesCategoryCatchData}
\alias{ResampleMeanSpeciesCategoryCatchData}
\title{Resamples biotic PSUs within Stratum in MeanSpeciesCategoryCatchData}
\usage{
ResampleMeanSpeciesCategoryCatchData(MeanSpeciesCategoryCatchData, Seed)
}
\arguments{
\item{MeanSpeciesCategoryCatchData}{The \code{\link[RstoxBase]{MeanSpeciesCategoryCatchData}} data.}

\item{Seed}{The seed, given as a sinigle initeger.}
}
\description{
This function resamples biotic PSUs with replacement within each Stratum, changing the MeanSpeciesCategoryCatchData
}
