% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{runFunction}
\alias{runFunction}
\alias{runFunction.JSON}
\title{General functions to run a function of an Rstox package, modeled by do.call().}
\usage{
runFunction(
  what,
  args,
  package = "RstoxFramework",
  removeCall = TRUE,
  onlyStoxMessages = TRUE
)

runFunction.JSON(cmd)
}
\arguments{
\item{what}{either a function or a non-empty character string naming the
    function to be called.}

\item{args}{a \emph{list} of arguments to the function call.  The
    \code{names} attribute of \code{args} gives the argument names.}

\item{package}{The name of the package holding the function given by \code{what}.}

\item{removeCall}{Logical: If FALSE, keep the call in the error message.}

\item{onlyStoxMessages}{Logical: If TRUE show only the StoX messages, which are those starting with "StoX: ".}

\item{cmd}{A JSON string containing parameters listed above.}
}
\description{
\code{runFunction} runs a function using the \code{\link[base]{do.call}} syntax, whereas \code{runFunction.JSON} accepts a JSON string conntaining the parameters to pass on to \code{runFunction}.
}
