% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{runProcess}
\alias{runProcess}
\title{Run one process}
\usage{
runProcess(
  projectPath,
  modelName,
  processID,
  msg = TRUE,
  save = TRUE,
  fileOutput = NULL,
  setUseProcessDataToTRUE = TRUE,
  purge.processData = FALSE,
  replaceArgs = list(),
  replaceData = NULL,
  output.file.type = c("text", "binary")
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}

\item{replaceData}{Either the data to replace the process output by, or a list of two elements \code{FunctionName} and \code{MoreArgs}, giving a function to apply to the output from the process with additional arguments stored in \code{MoreArgs}. The function in applied using \code{\link{do.call}}, with \code{args} being a list with the process output first, followed by the \code{MoreArgs}.}
}
\description{
Run one process
}
