% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{runProcesses}
\alias{runProcesses}
\title{Run processes of a model.}
\usage{
runProcesses(
  projectPath,
  modelName,
  startProcess = 1,
  endProcess = Inf,
  msg = TRUE,
  save = TRUE,
  force.restart = FALSE,
  fileOutput = NULL,
  setUseProcessDataToTRUE = TRUE,
  purge.processData = FALSE,
  replaceDataList = list(),
  replaceArgs = list(),
  output.file.type = c("default", "text", "RData", "rds"),
  ...
)
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (one of "baseline", "analysis" and "report").}

\item{replaceDataList}{A list named by the processes to replace output data for. See \code{\link{runProcess}}.}
}
\description{
Run processes of a model.
}
