% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{unlistToDataType}
\alias{unlistToDataType}
\title{Function to flatten a list of output datat and add names from the levels of the list}
\usage{
unlistToDataType(
  processOutput,
  sep = "_",
  validOutputDataClasses = getRstoxFrameworkDefinitions("validOutputDataClasses"),
  nlevel = 2,
  keepNonStandardAttributes = FALSE
)
}
\arguments{
\item{processOutput}{A list of StoX output data.}

\item{sep}{The separator to use when constructing names for the final flat list, defaulted to underscore, but slash can also be useful e.g. for denoting groupes in a NetCDF4 file.}

\item{validOutputDataClasses}{A vector of valid output data classes, indicating when to stop the unlisting.}

\item{nlevel}{The number of levels to unlist through.}

\item{keepNonStandardAttributes}{Logical: If TRUE, keep attributes other than the standard \code{dim}, \code{names} and \code{dimnames}.}
}
\description{
Function to flatten a list of output datat and add names from the levels of the list
}
