% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{unzipProject}
\alias{unzipProject}
\title{Function for unzipping a zipped StoX project}
\usage{
unzipProject(projectPath, exdir = ".")
}
\arguments{
\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{exdir}{The directory to unzip to, defaulted to the current directory.}
}
\description{
Function for unzipping a zipped StoX project
}
