% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framework.R
\name{writeStoxOutput}
\alias{writeStoxOutput}
\title{Write the output of a function}
\usage{
writeStoxOutput(
  x,
  folderPath,
  filePath,
  ow = FALSE,
  escape = TRUE,
  projectPath = projectPath,
  modelName = modelName,
  processID = processID
)
}
\arguments{
\item{x}{The output from a StoX function.}

\item{folderPath}{Character: The path to the folder to write the files to}

\item{filePath}{Character: Optional. If given this overrides the default file paths. Must be of the same length as \code{x} is \code{x} is a list of tables etc.}

\item{ow}{Logical: If TRUE overwrite the file if existing.}

\item{escape}{Logical: Should strings be escaped in text output. Defaulted to TRUE.}

\item{projectPath}{The path to the StoX project, i.e., the folder of the project with the sub folders "input", "output" and "process". Can possibly be the path to a file inside the project folder.}

\item{modelName}{The name of the model (possible values are "baseline", "analysis" and "report").}

\item{processID}{The ID of the process.}
}
\description{
Write the output of a function
}
